Network.AWS.CloudWatch

Description

Amazon CloudWatch is a monitoring service for AWS cloud resources and the applications you run on AWS. You can use Amazon CloudWatch to collect and track metrics, collect and monitor log files, and set alarms. Amazon CloudWatch can monitor AWS resources such as Amazon EC2 instances, Amazon DynamoDB tables, and Amazon RDS DB instances, as well as custom metrics generated by your applications and services, and any log files your applications generate. You can use Amazon CloudWatch to gain system-wide visibility into resource utilization, application performance, and operational health. You can use these insights to react and keep your application running smoothly.

Documentation

module Network.AWS.CloudWatch.DeleteAlarms

module Network.AWS.CloudWatch.DescribeAlarmHistory

module Network.AWS.CloudWatch.DescribeAlarms

module Network.AWS.CloudWatch.DescribeAlarmsForMetric

module Network.AWS.CloudWatch.DisableAlarmActions

module Network.AWS.CloudWatch.EnableAlarmActions

module Network.AWS.CloudWatch.GetMetricStatistics

module Network.AWS.CloudWatch.ListMetrics

module Network.AWS.CloudWatch.PutMetricAlarm

module Network.AWS.CloudWatch.PutMetricData

module Network.AWS.CloudWatch.SetAlarmState

module Network.AWS.CloudWatch.Types