Network.AWS.CloudWatch.DeleteAlarms

Description

Deletes the specified alarms. In the event of an error, no alarms are deleted.

Creating a Request

deleteAlarms :: DeleteAlarms

Creates a value of DeleteAlarms with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dAlarmNames - The alarms to be deleted.

data DeleteAlarms

See: deleteAlarms smart constructor.

Request Lenses

dAlarmNames :: Lens' DeleteAlarms [Text]

The alarms to be deleted.

Destructuring the Response

deleteAlarmsResponse :: DeleteAlarmsResponse

Creates a value of DeleteAlarmsResponse with the minimum fields required to make a request.

data DeleteAlarmsResponse

See: deleteAlarmsResponse smart constructor.