Network.AWS.CloudWatch.DescribeAlarmHistory

Description

Retrieves history for the specified alarm. Filter alarms by date range or item type. If an alarm name is not specified, Amazon CloudWatch returns histories for all of the owner's alarms. Amazon CloudWatch retains the history of an alarm for two weeks, whether or not you delete the alarm.

Request

data DescribeAlarmHistory

Request constructor

describeAlarmHistory :: DescribeAlarmHistory

Smart constructor for the minimum required parameters to construct a valid DescribeAlarmHistory request.

The fields accessible through corresponding lenses are:

AlarmName :: Maybe Text

HistoryItemType :: Maybe HistoryItemType

StartDate :: Maybe ISO8601

EndDate :: Maybe ISO8601

MaxRecords :: Maybe Integer

NextToken :: Maybe Text

Request lenses

dahAlarmName :: Lens' DescribeAlarmHistory (Maybe Text)

The name of the alarm.

dahHistoryItemType :: Lens' DescribeAlarmHistory (Maybe HistoryItemType)

The type of alarm histories to retrieve.

dahStartDate :: Lens' DescribeAlarmHistory (Maybe ISO8601)

The starting date to retrieve alarm history.

dahEndDate :: Lens' DescribeAlarmHistory (Maybe ISO8601)

The ending date to retrieve alarm history.

dahMaxRecords :: Lens' DescribeAlarmHistory (Maybe Integer)

The maximum number of alarm history records to retrieve.

dahNextToken :: Lens' DescribeAlarmHistory (Maybe Text)

The token returned by a previous call to indicate that there is more data available.

Response

data DescribeAlarmHistoryResponse

The output for the DescribeAlarmHistory action.

Response constructor

describeAlarmHistoryResponse :: DescribeAlarmHistoryResponse

Smart constructor for the minimum required parameters to construct a valid DescribeAlarmHistoryResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

AlarmHistoryItems :: [AlarmHistoryItem]

NextToken :: Maybe Text

Response lenses

dahrAlarmHistoryItems :: Lens' DescribeAlarmHistoryResponse [AlarmHistoryItem]

A list of alarm histories in JSON format.

dahrNextToken :: Lens' DescribeAlarmHistoryResponse (Maybe Text)

A string that marks the start of the next batch of returned results.