Network.AWS.CloudWatch.DescribeAlarms

Description

Retrieves alarms with the specified names. If no name is specified, all alarms for the user are returned. Alarms can be retrieved by using only a prefix for the alarm name, the alarm state, or a prefix for any action.

Request

data DescribeAlarms

Request constructor

describeAlarms :: DescribeAlarms

Smart constructor for the minimum required parameters to construct a valid DescribeAlarms request.

The fields accessible through corresponding lenses are:

AlarmNames :: [Text]

AlarmNamePrefix :: Maybe Text

StateValue :: Maybe StateValue

ActionPrefix :: Maybe Text

MaxRecords :: Maybe Integer

NextToken :: Maybe Text

Request lenses

da1AlarmNames :: Lens' DescribeAlarms [Text]

A list of alarm names to retrieve information for.

da1AlarmNamePrefix :: Lens' DescribeAlarms (Maybe Text)

The alarm name prefix. AlarmNames cannot be specified if this parameter is specified.

da1StateValue :: Lens' DescribeAlarms (Maybe StateValue)

The state value to be used in matching alarms.

da1ActionPrefix :: Lens' DescribeAlarms (Maybe Text)

The action name prefix.

da1MaxRecords :: Lens' DescribeAlarms (Maybe Integer)

The maximum number of alarm descriptions to retrieve.

da1NextToken :: Lens' DescribeAlarms (Maybe Text)

The token returned by a previous call to indicate that there is more data available.

Response

data DescribeAlarmsResponse

The output for the DescribeAlarms action.

Response constructor

describeAlarmsResponse :: DescribeAlarmsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeAlarmsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

MetricAlarms :: [MetricAlarm]

NextToken :: Maybe Text

Response lenses

darMetricAlarms :: Lens' DescribeAlarmsResponse [MetricAlarm]

A list of information for the specified alarms.

darNextToken :: Lens' DescribeAlarmsResponse (Maybe Text)

A string that marks the start of the next batch of returned results.