Network.AWS.CloudWatch.DescribeAlarmsForMetric

Description

Retrieves all alarms for a single metric. Specify a statistic, period, or unit to filter the set of alarms further.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_DescribeAlarmsForMetric.html

Request

data DescribeAlarmsForMetric

Request constructor

describeAlarmsForMetric

Arguments

:: Text
dafmMetricName

-> Text
dafmNamespace

-> DescribeAlarmsForMetric

DescribeAlarmsForMetric constructor.

The fields accessible through corresponding lenses are:

dafmDimensions :: [Dimension]

dafmMetricName :: Text

dafmNamespace :: Text

dafmPeriod :: Maybe Natural

dafmStatistic :: Maybe Statistic

dafmUnit :: Maybe StandardUnit

Request lenses

dafmDimensions :: Lens' DescribeAlarmsForMetric [Dimension]

The list of dimensions associated with the metric.

dafmMetricName :: Lens' DescribeAlarmsForMetric Text

The name of the metric.

dafmNamespace :: Lens' DescribeAlarmsForMetric Text

The namespace of the metric.

dafmPeriod :: Lens' DescribeAlarmsForMetric (Maybe Natural)

The period in seconds over which the statistic is applied.

dafmStatistic :: Lens' DescribeAlarmsForMetric (Maybe Statistic)

The statistic for the metric.

dafmUnit :: Lens' DescribeAlarmsForMetric (Maybe StandardUnit)

The unit for the metric.

Response

data DescribeAlarmsForMetricResponse

Response constructor

describeAlarmsForMetricResponse :: DescribeAlarmsForMetricResponse

DescribeAlarmsForMetricResponse constructor.

The fields accessible through corresponding lenses are:

dafmrMetricAlarms :: [MetricAlarm]

Response lenses

dafmrMetricAlarms :: Lens' DescribeAlarmsForMetricResponse [MetricAlarm]

A list of information for each alarm with the specified metric.