Network.AWS.CloudWatch.DisableAlarmActions

Description

Disables actions for the specified alarms. When an alarm's actions are disabled the alarm's state may change, but none of the alarm's actions will execute.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_DisableAlarmActions.html

Request

data DisableAlarmActions

Request constructor

disableAlarmActions :: DisableAlarmActions

DisableAlarmActions constructor.

The fields accessible through corresponding lenses are:

daaAlarmNames :: [Text]

Request lenses

daaAlarmNames :: Lens' DisableAlarmActions [Text]

The names of the alarms to disable actions for.

Response

data DisableAlarmActionsResponse

Response constructor

disableAlarmActionsResponse :: DisableAlarmActionsResponse

DisableAlarmActionsResponse constructor.