Network.AWS.CloudWatch.GetMetricStatistics

Description

Gets statistics for the specified metric. The maximum number of data points returned from a single GetMetricStatistics request is 1,440. If a request is made that generates more than 1,440 data points, Amazon CloudWatch returns an error. In such a case, alter the request by narrowing the specified time range or increasing the specified period. Alternatively, make multiple requests across adjacent time ranges. Amazon CloudWatch aggregates data points based on the length of the period that you specify. For example, if you request statistics with a one-minute granularity, Amazon CloudWatch aggregates data points with time stamps that fall within the same one-minute period. In such a case, the data points queried can greatly outnumber the data points returned. The maximum number of data points that can be queried is 50,850; whereas the maximum number of data points returned is 1,440. The following examples show various statistics allowed by the data point query maximum of 50,850 when you call GetMetricStatistics on Amazon EC2 instances with detailed (one-minute) monitoring enabled: Statistics for up to 400 instances for a span of one hour Statistics for up to 35 instances over a span of 24 hours Statistics for up to 2 instances over a span of 2 weeks.

Request

data GetMetricStatistics

Request constructor

mkGetMetricStatistics

Arguments

:: Text
gmsNamespace

-> Text
gmsMetricName

-> ISO8601
gmsStartTime

-> ISO8601
gmsEndTime

-> Integer
gmsPeriod

-> List1 Statistic
gmsStatistics

-> GetMetricStatistics

Smart constructor for the minimum required parameters to construct a valid GetMetricStatistics request.

The fields accessible through corresponding lenses are:

Namespace :: Text

MetricName :: Text

Dimensions :: [Dimension]

StartTime :: ISO8601

EndTime :: ISO8601

Period :: Integer

Statistics :: List1 Statistic

Unit :: Maybe StandardUnit

Request lenses

gmsNamespace :: Lens' GetMetricStatistics Text

The namespace of the metric.

gmsMetricName :: Lens' GetMetricStatistics Text

The name of the metric.

gmsDimensions :: Lens' GetMetricStatistics [Dimension]

A list of dimensions describing qualities of the metric.

gmsStartTime :: Lens' GetMetricStatistics ISO8601

The time stamp to use for determining the first datapoint to return. The value specified is inclusive; results include datapoints with the time stamp specified. The specified start time is rounded down to the nearest value. Datapoints are returned for start times up to two weeks in the past. Specified start times that are more than two weeks in the past will not return datapoints for metrics that are older than two weeks.

gmsEndTime :: Lens' GetMetricStatistics ISO8601

The time stamp to use for determining the last datapoint to return. The value specified is exclusive; results will include datapoints up to the time stamp specified.

gmsPeriod :: Lens' GetMetricStatistics Integer

The granularity, in seconds, of the returned datapoints. Period must be at least 60 seconds and must be a multiple of 60. The default value is 60.

gmsStatistics :: Lens' GetMetricStatistics (List1 Statistic)

The metric statistics to return.

gmsUnit :: Lens' GetMetricStatistics (Maybe StandardUnit)

The unit for the metric.

Response

data GetMetricStatisticsResponse

The output for the GetMetricStatistics action.

Response constructor

mkGetMetricStatisticsResponse :: GetMetricStatisticsResponse

Smart constructor for the minimum required parameters to construct a valid GetMetricStatisticsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Label :: Maybe Text

Datapoints :: [Datapoint]

Response lenses

gmsrLabel :: Lens' GetMetricStatisticsResponse (Maybe Text)

A label describing the specified metric.

gmsrDatapoints :: Lens' GetMetricStatisticsResponse [Datapoint]

The datapoints for the specified metric.