Network.AWS.CloudWatch.ListMetrics

Description

Returns a list of valid metrics stored for the AWS account owner. Returned metrics can be used with GetMetricStatistics to obtain statistical data for a given metric.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_ListMetrics.html

Request

data ListMetrics

Request constructor

listMetrics :: ListMetrics

ListMetrics constructor.

The fields accessible through corresponding lenses are:

lmDimensions :: [DimensionFilter]

lmMetricName :: Maybe Text

lmNamespace :: Maybe Text

lmNextToken :: Maybe Text

Request lenses

lmDimensions :: Lens' ListMetrics [DimensionFilter]

A list of dimensions to filter against.

lmMetricName :: Lens' ListMetrics (Maybe Text)

The name of the metric to filter against.

lmNamespace :: Lens' ListMetrics (Maybe Text)

The namespace to filter against.

lmNextToken :: Lens' ListMetrics (Maybe Text)

The token returned by a previous call to indicate that there is more data available.

Response

data ListMetricsResponse

Response constructor

listMetricsResponse :: ListMetricsResponse

ListMetricsResponse constructor.

The fields accessible through corresponding lenses are:

lmrMetrics :: [Metric]

lmrNextToken :: Maybe Text

Response lenses

lmrMetrics :: Lens' ListMetricsResponse [Metric]

A list of metrics used to generate statistics for an AWS account.

lmrNextToken :: Lens' ListMetricsResponse (Maybe Text)

A string that marks the start of the next batch of returned results.