Network.AWS.CloudWatch.Monadic

Description

Amazon CloudWatch is a monitoring service for AWS cloud resources and the applications you run on AWS. You can use Amazon CloudWatch to collect and track metrics, collect and monitor log files, and set alarms. Amazon CloudWatch can monitor AWS resources such as Amazon EC2 instances, Amazon DynamoDB tables, and Amazon RDS DB instances, as well as custom metrics generated by your applications and services, and any log files your applications generate. You can use Amazon CloudWatch to gain system-wide visibility into resource utilization, application performance, and operational health. You can use these insights to react and keep your application running smoothly.

This module is provided for convenience. It offers an alternative to the common idiom of supplying required fields to an operations's smart constructor, using the operation's lenses to modify additional fields, and then sending or paginating the request.

As an example: using Network.AWS.CloudWatch with the smart constructor and basic lens syntax, before explicitly calling send:

import Control.Monad.Trans.AWS
import Network.AWS.CloudWatch

send $ (mkOperationName w x)
     & onLensField1 .~ y
     & onLensField2 .~ z

Versus using Network.AWS.CloudWatch.Monadic with the State operator variants from Control.Lens.Setter such as .= to modify any additional request parameters before sending:

import Network.AWS.CloudWatch.Monadic

operationName w x $ do
    onLensField1 .= y
    onLensField2 .= z

DeleteAlarms

Deletes all specified alarms. In the event of an error, no alarms are deleted.

See: CloudWatch

deleteAlarms

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> [Text]
daAlarmNames

-> State DeleteAlarms a

-> m DeleteAlarmsResponse

deleteAlarmsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> [Text]
daAlarmNames

-> State DeleteAlarms a

-> m (Either ServiceEr DeleteAlarmsResponse)

DescribeAlarmHistory

Retrieves history for the specified alarm. Filter alarms by date range or item type. If an alarm name is not specified, Amazon CloudWatch returns histories for all of the owner's alarms. Amazon CloudWatch retains the history of an alarm for two weeks, whether or not you delete the alarm.

See: CloudWatch

describeAlarmHistory :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env (ResumableSource m)) => State DescribeAlarmHistory a -> ResumableSource m DescribeAlarmHistoryResponse

describeAlarmHistoryCatch :: (MonadCatch m, MonadResource m, MonadReader Env (ResumableSource m)) => State DescribeAlarmHistory a -> ResumableSource m (Either ServiceEr DescribeAlarmHistoryResponse)

DescribeAlarms

Retrieves alarms with the specified names. If no name is specified, all alarms for the user are returned. Alarms can be retrieved by using only a prefix for the alarm name, the alarm state, or a prefix for any action.

See: CloudWatch

describeAlarms :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env (ResumableSource m)) => State DescribeAlarms a -> ResumableSource m DescribeAlarmsResponse

describeAlarmsCatch :: (MonadCatch m, MonadResource m, MonadReader Env (ResumableSource m)) => State DescribeAlarms a -> ResumableSource m (Either ServiceEr DescribeAlarmsResponse)

DescribeAlarmsForMetric

Retrieves all alarms for a single metric. Specify a statistic, period, or unit to filter the set of alarms further.

See: CloudWatch

describeAlarmsForMetric

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dafmMetricName

-> Text
dafmNamespace

-> State DescribeAlarmsForMetric a

-> m DescribeAlarmsForMetricResponse

describeAlarmsForMetricCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dafmMetricName

-> Text
dafmNamespace

-> State DescribeAlarmsForMetric a

-> m (Either ServiceEr DescribeAlarmsForMetricResponse)

DisableAlarmActions

Disables actions for the specified alarms. When an alarm's actions are disabled the alarm's state may change, but none of the alarm's actions will execute.

See: CloudWatch

disableAlarmActions

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> [Text]
daaAlarmNames

-> State DisableAlarmActions a

-> m DisableAlarmActionsResponse

disableAlarmActionsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> [Text]
daaAlarmNames

-> State DisableAlarmActions a

-> m (Either ServiceEr DisableAlarmActionsResponse)

EnableAlarmActions

Enables actions for the specified alarms.

See: CloudWatch

enableAlarmActions

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> [Text]
eaaAlarmNames

-> State EnableAlarmActions a

-> m EnableAlarmActionsResponse

enableAlarmActionsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> [Text]
eaaAlarmNames

-> State EnableAlarmActions a

-> m (Either ServiceEr EnableAlarmActionsResponse)

GetMetricStatistics

Gets statistics for the specified metric. The maximum number of data points returned from a single GetMetricStatistics request is 1,440. If a request is made that generates more than 1,440 data points, Amazon CloudWatch returns an error. In such a case, alter the request by narrowing the specified time range or increasing the specified period. Alternatively, make multiple requests across adjacent time ranges. Amazon CloudWatch aggregates data points based on the length of the period that you specify. For example, if you request statistics with a one-minute granularity, Amazon CloudWatch aggregates data points with time stamps that fall within the same one-minute period. In such a case, the data points queried can greatly outnumber the data points returned. The maximum number of data points that can be queried is 50,850; whereas the maximum number of data points returned is 1,440. The following examples show various statistics allowed by the data point query maximum of 50,850 when you call GetMetricStatistics on Amazon EC2 instances with detailed (one-minute) monitoring enabled: Statistics for up to 400 instances for a span of one hour Statistics for up to 35 instances over a span of 24 hours Statistics for up to 2 instances over a span of 2 weeks.

See: CloudWatch

getMetricStatistics

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
gmsNamespace

-> Text
gmsMetricName

-> ISO8601
gmsStartTime

-> ISO8601
gmsEndTime

-> Integer
gmsPeriod

-> List1 Statistic
gmsStatistics

-> State GetMetricStatistics a

-> m GetMetricStatisticsResponse

getMetricStatisticsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
gmsNamespace

-> Text
gmsMetricName

-> ISO8601
gmsStartTime

-> ISO8601
gmsEndTime

-> Integer
gmsPeriod

-> List1 Statistic
gmsStatistics

-> State GetMetricStatistics a

-> m (Either ServiceEr GetMetricStatisticsResponse)

ListMetrics

Returns a list of valid metrics stored for the AWS account owner. Returned metrics can be used with GetMetricStatistics to obtain statistical data for a given metric. Up to 500 results are returned for any one call. To retrieve further results, use returned NextToken values with subsequent ListMetrics operations. If you create a metric with the PutMetricData action, allow up to fifteen minutes for the metric to appear in calls to the ListMetrics action.

See: CloudWatch

listMetrics :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env (ResumableSource m)) => State ListMetrics a -> ResumableSource m ListMetricsResponse

listMetricsCatch :: (MonadCatch m, MonadResource m, MonadReader Env (ResumableSource m)) => State ListMetrics a -> ResumableSource m (Either ServiceEr ListMetricsResponse)

PutMetricAlarm

Creates or updates an alarm and associates it with the specified Amazon CloudWatch metric. Optionally, this operation can associate one or more Amazon Simple Notification Service resources with the alarm. When this operation creates an alarm, the alarm state is immediately set to INSUFFICIENT_DATA. The alarm is evaluated and its StateValue is set appropriately. Any actions associated with the StateValue is then executed. When updating an existing alarm, its StateValue is left unchanged.

See: CloudWatch

putMetricAlarm

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
pmaAlarmName

-> Integer
pmaPeriod

-> Integer
pmaEvaluationPeriods

-> Double
pmaThreshold

-> ComparisonOperator
pmaComparisonOperator

-> Text
pmaMetricName

-> Text
pmaNamespace

-> Statistic
pmaStatistic

-> State PutMetricAlarm a

-> m PutMetricAlarmResponse

putMetricAlarmCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
pmaAlarmName

-> Integer
pmaPeriod

-> Integer
pmaEvaluationPeriods

-> Double
pmaThreshold

-> ComparisonOperator
pmaComparisonOperator

-> Text
pmaMetricName

-> Text
pmaNamespace

-> Statistic
pmaStatistic

-> State PutMetricAlarm a

-> m (Either ServiceEr PutMetricAlarmResponse)

PutMetricData

Publishes metric data points to Amazon CloudWatch. Amazon Cloudwatch associates the data points with the specified metric. If the specified metric does not exist, Amazon CloudWatch creates the metric. If you create a metric with the PutMetricData action, allow up to fifteen minutes for the metric to appear in calls to the ListMetrics action. The size of a PutMetricData request is limited to 8 KB for HTTP GET requests and 40 KB for HTTP POST requests. Although the Value parameter accepts numbers of type Double, Amazon CloudWatch truncates values with very large exponents. Values with base-10 exponents greater than 126 (1 x 10^126) are truncated. Likewise, values with base-10 exponents less than -130 (1 x 10^-130) are also truncated.

See: CloudWatch

putMetricData

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
pmdNamespace

-> [MetricDatum]
pmdMetricData

-> State PutMetricData a

-> m PutMetricDataResponse

putMetricDataCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
pmdNamespace

-> [MetricDatum]
pmdMetricData

-> State PutMetricData a

-> m (Either ServiceEr PutMetricDataResponse)

SetAlarmState

Temporarily sets the state of an alarm. When the updated StateValue differs from the previous value, the action configured for the appropriate state is invoked. This is not a permanent change. The next periodic alarm check (in about a minute) will set the alarm to its actual state.

See: CloudWatch

setAlarmState

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
sasAlarmName

-> StateValue
sasStateValue

-> Text
sasStateReason

-> State SetAlarmState a

-> m SetAlarmStateResponse

setAlarmStateCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
sasAlarmName

-> StateValue
sasStateValue

-> Text
sasStateReason

-> State SetAlarmState a

-> m (Either ServiceEr SetAlarmStateResponse)

Re-exported

module Network.AWS.CloudWatch

(.=) :: MonadState s m => ASetter s s a b -> b -> m ()

(?=) :: MonadState s m => ASetter s s a (Maybe b) -> b -> m ()

(<>=) :: (MonadState s m, Monoid a) => ASetter' s a -> a -> m ()

(%=) :: (Profunctor p, MonadState s m) => Setting p s s a b -> p a b -> m ()