Network.AWS.CloudWatch.PutDashboard

Description

Creates a dashboard if it does not already exist, or updates an existing dashboard. If you update a dashboard, the entire contents are replaced with what you specify here.

You can have up to 500 dashboards per account. All dashboards in your account are global, not region-specific.

A simple way to create a dashboard using PutDashboard is to copy an existing dashboard. To copy an existing dashboard using the console, you can load the dashboard and then use the View/edit source command in the Actions menu to display the JSON block for that dashboard. Another way to copy a dashboard is to use GetDashboard , and then use the data returned within DashboardBody as the template for the new dashboard when you call PutDashboard .

When you create a dashboard with PutDashboard , a good practice is to add a text widget at the top of the dashboard with a message that the dashboard was created by script and should not be changed in the console. This message could also point console users to the location of the DashboardBody script or the CloudFormation template used to create the dashboard.

Creating a Request

putDashboard :: PutDashboard

Creates a value of PutDashboard with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pdDashboardName - The name of the dashboard. If a dashboard with this name already exists, this call modifies that dashboard, replacing its current contents. Otherwise, a new dashboard is created. The maximum length is 255, and valid characters are A-Z, a-z, 0-9, "-", and "_".

pdDashboardBody - The detailed information about the dashboard in JSON format, including the widgets to include and their location on the dashboard. For more information about the syntax, see 'CloudWatch-Dashboard-Body-Structure' .

data PutDashboard

See: putDashboard smart constructor.

Request Lenses

pdDashboardName :: Lens' PutDashboard (Maybe Text)

The name of the dashboard. If a dashboard with this name already exists, this call modifies that dashboard, replacing its current contents. Otherwise, a new dashboard is created. The maximum length is 255, and valid characters are A-Z, a-z, 0-9, "-", and "_".

pdDashboardBody :: Lens' PutDashboard (Maybe Text)

The detailed information about the dashboard in JSON format, including the widgets to include and their location on the dashboard. For more information about the syntax, see 'CloudWatch-Dashboard-Body-Structure' .

Destructuring the Response

putDashboardResponse

Arguments

:: Int
pdrsResponseStatus

-> PutDashboardResponse

Creates a value of PutDashboardResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pdrsDashboardValidationMessages - If the input for PutDashboard was correct and the dashboard was successfully created or modified, this result is empty. If this result includes only warning messages, then the input was valid enough for the dashboard to be created or modified, but some elements of the dashboard may not render. If this result includes error messages, the input was not valid and the operation failed.

pdrsResponseStatus - -- | The response status code.

data PutDashboardResponse

See: putDashboardResponse smart constructor.

Response Lenses

pdrsDashboardValidationMessages :: Lens' PutDashboardResponse [DashboardValidationMessage]

If the input for PutDashboard was correct and the dashboard was successfully created or modified, this result is empty. If this result includes only warning messages, then the input was valid enough for the dashboard to be created or modified, but some elements of the dashboard may not render. If this result includes error messages, the input was not valid and the operation failed.

pdrsResponseStatus :: Lens' PutDashboardResponse Int

- | The response status code.