Network.AWS.CloudWatch.PutMetricAlarm

Description

Creates or updates an alarm and associates it with the specified Amazon CloudWatch metric. Optionally, this operation can associate one or more Amazon Simple Notification Service resources with the alarm.

When this operation creates an alarm, the alarm state is immediately set to INSUFFICIENT_DATA. The alarm is evaluated and its StateValue is set appropriately. Any actions associated with the StateValue is then executed.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_PutMetricAlarm.html

Request

data PutMetricAlarm

Request constructor

putMetricAlarm

Arguments

:: Text
pmaAlarmName

-> Text
pmaMetricName

-> Text
pmaNamespace

-> Statistic
pmaStatistic

-> Natural
pmaPeriod

-> Natural
pmaEvaluationPeriods

-> Double
pmaThreshold

-> ComparisonOperator
pmaComparisonOperator

-> PutMetricAlarm

PutMetricAlarm constructor.

The fields accessible through corresponding lenses are:

pmaActionsEnabled :: Maybe Bool

pmaAlarmActions :: [Text]

pmaAlarmDescription :: Maybe Text

pmaAlarmName :: Text

pmaComparisonOperator :: ComparisonOperator

pmaDimensions :: [Dimension]

pmaEvaluationPeriods :: Natural

pmaInsufficientDataActions :: [Text]

pmaMetricName :: Text

pmaNamespace :: Text

pmaOKActions :: [Text]

pmaPeriod :: Natural

pmaStatistic :: Statistic

pmaThreshold :: Double

pmaUnit :: Maybe StandardUnit

Request lenses

pmaActionsEnabled :: Lens' PutMetricAlarm (Maybe Bool)

Indicates whether or not actions should be executed during any changes to the alarm's state.

pmaAlarmActions :: Lens' PutMetricAlarm [Text]

The list of actions to execute when this alarm transitions into an ALARM state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only action supported is publishing to an Amazon SNS topic or an Amazon Auto Scaling policy.

pmaAlarmDescription :: Lens' PutMetricAlarm (Maybe Text)

The description for the alarm.

pmaAlarmName :: Lens' PutMetricAlarm Text

The descriptive name for the alarm. This name must be unique within the user's AWS account

pmaComparisonOperator :: Lens' PutMetricAlarm ComparisonOperator

The arithmetic operation to use when comparing the specified Statistic and Threshold. The specified Statistic value is used as the first operand.

pmaDimensions :: Lens' PutMetricAlarm [Dimension]

The dimensions for the alarm's associated metric.

pmaEvaluationPeriods :: Lens' PutMetricAlarm Natural

The number of periods over which data is compared to the specified threshold.

pmaInsufficientDataActions :: Lens' PutMetricAlarm [Text]

The list of actions to execute when this alarm transitions into an INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only action supported is publishing to an Amazon SNS topic or an Amazon Auto Scaling policy.

pmaMetricName :: Lens' PutMetricAlarm Text

The name for the alarm's associated metric.

pmaNamespace :: Lens' PutMetricAlarm Text

The namespace for the alarm's associated metric.

pmaOKActions :: Lens' PutMetricAlarm [Text]

The list of actions to execute when this alarm transitions into an OK state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only action supported is publishing to an Amazon SNS topic or an Amazon Auto Scaling policy.

pmaPeriod :: Lens' PutMetricAlarm Natural

The period in seconds over which the specified statistic is applied.

pmaStatistic :: Lens' PutMetricAlarm Statistic

The statistic to apply to the alarm's associated metric.

pmaThreshold :: Lens' PutMetricAlarm Double

The value against which the specified statistic is compared.

pmaUnit :: Lens' PutMetricAlarm (Maybe StandardUnit)

The unit for the alarm's associated metric.

Response

data PutMetricAlarmResponse

Response constructor

putMetricAlarmResponse :: PutMetricAlarmResponse

PutMetricAlarmResponse constructor.