Network.AWS.CloudWatch.PutMetricData

Description

Publishes metric data points to Amazon CloudWatch. Amazon Cloudwatch associates the data points with the specified metric. If the specified metric does not exist, Amazon CloudWatch creates the metric. It can take up to fifteen minutes for a new metric to appear in calls to the ListMetrics action.

The size of a PutMetricData request is limited to 8 KB for HTTP GET requests and 40 KB for HTTP POST requests.

Although the Value parameter accepts numbers of type Double, Amazon CloudWatch truncates values with very large exponents. Values with base-10 exponents greater than 126 (1 x 10^126) are truncated. Likewise, values with base-10 exponents less than -130 (1 x 10^-130) are also truncated. Data that is timestamped 24 hours or more in the past may take in excess of 48 hours to become available from submission time using GetMetricStatistics.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_PutMetricData.html

Request

data PutMetricData

Request constructor

putMetricData

Arguments

:: Text
pmdNamespace

-> PutMetricData

PutMetricData constructor.

The fields accessible through corresponding lenses are:

pmdMetricData :: [MetricDatum]

pmdNamespace :: Text

Request lenses

pmdMetricData :: Lens' PutMetricData [MetricDatum]

A list of data describing the metric.

pmdNamespace :: Lens' PutMetricData Text

The namespace for the metric data.

Response

data PutMetricDataResponse

Response constructor

putMetricDataResponse :: PutMetricDataResponse

PutMetricDataResponse constructor.