Network.AWS.CloudWatch.SetAlarmState

Description

Temporarily sets the state of an alarm. When the updated StateValue differs from the previous value, the action configured for the appropriate state is invoked. This is not a permanent change. The next periodic alarm check (in about a minute) will set the alarm to its actual state.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_SetAlarmState.html

Request

data SetAlarmState

Request constructor

setAlarmState

Arguments

:: Text
sasAlarmName

-> StateValue
sasStateValue

-> Text
sasStateReason

-> SetAlarmState

SetAlarmState constructor.

The fields accessible through corresponding lenses are:

sasAlarmName :: Text

sasStateReason :: Text

sasStateReasonData :: Maybe Text

sasStateValue :: StateValue

Request lenses

sasAlarmName :: Lens' SetAlarmState Text

The descriptive name for the alarm. This name must be unique within the user's AWS account. The maximum length is 255 characters.

sasStateReason :: Lens' SetAlarmState Text

The reason that this alarm is set to this specific state (in human-readable text format)

sasStateReasonData :: Lens' SetAlarmState (Maybe Text)

The reason that this alarm is set to this specific state (in machine-readable JSON format)

sasStateValue :: Lens' SetAlarmState StateValue

The value of the state.

Response

data SetAlarmStateResponse

Response constructor

setAlarmStateResponse :: SetAlarmStateResponse

SetAlarmStateResponse constructor.