Network.AWS.CloudWatch.Types

Description

Amazon CloudWatch is a monitoring service for AWS cloud resources and the applications you run on AWS. You can use Amazon CloudWatch to collect and track metrics, collect and monitor log files, and set alarms. Amazon CloudWatch can monitor AWS resources such as Amazon EC2 instances, Amazon DynamoDB tables, and Amazon RDS DB instances, as well as custom metrics generated by your applications and services, and any log files your applications generate. You can use Amazon CloudWatch to gain system-wide visibility into resource utilization, application performance, and operational health. You can use these insights to react and keep your application running smoothly.

Service

data CloudWatch

Supported version (2010-08-01) of the Amazon CloudWatch service.

XML

xmlOptions :: Tagged a XMLOptions

ComparisonOperator

data ComparisonOperator

Constructors

ComparisonOperatorGreaterThanOrEqualToThreshold
GreaterThanOrEqualToThreshold

ComparisonOperatorGreaterThanThreshold
GreaterThanThreshold

ComparisonOperatorLessThanOrEqualToThreshold
LessThanOrEqualToThreshold

ComparisonOperatorLessThanThreshold
LessThanThreshold

HistoryItemType

data HistoryItemType

Constructors

HistoryItemTypeAction
Action

HistoryItemTypeConfigurationUpdate
ConfigurationUpdate

HistoryItemTypeStateUpdate
StateUpdate

StandardUnit

data StandardUnit

Constructors

StandardUnitBits
Bits

StandardUnitBitsSecond
Bits/Second

StandardUnitBytes
Bytes

StandardUnitBytesSecond
Bytes/Second

StandardUnitCount
Count

StandardUnitCountSecond
Count/Second

StandardUnitGigabits
Gigabits

StandardUnitGigabitsSecond
Gigabits/Second

StandardUnitGigabytes
Gigabytes

StandardUnitGigabytesSecond
Gigabytes/Second

StandardUnitKilobits
Kilobits

StandardUnitKilobitsSecond
Kilobits/Second

StandardUnitKilobytes
Kilobytes

StandardUnitKilobytesSecond
Kilobytes/Second

StandardUnitMegabits
Megabits

StandardUnitMegabitsSecond
Megabits/Second

StandardUnitMegabytes
Megabytes

StandardUnitMegabytesSecond
Megabytes/Second

StandardUnitMicroseconds
Microseconds

StandardUnitMilliseconds
Milliseconds

StandardUnitNone
None

StandardUnitPercent
Percent

StandardUnitSeconds
Seconds

StandardUnitTerabits
Terabits

StandardUnitTerabitsSecond
Terabits/Second

StandardUnitTerabytes
Terabytes

StandardUnitTerabytesSecond
Terabytes/Second

StateValue

data StateValue

Constructors

StateValueAlarm
ALARM

StateValueInsufficientData
INSUFFICIENT_DATA

StateValueOk
OK

Statistic

data Statistic

Constructors

StatisticAverage
Average

StatisticMaximum
Maximum

StatisticMinimum
Minimum

StatisticSampleCount
SampleCount

StatisticSum
Sum

AlarmHistoryItem

data AlarmHistoryItem

The AlarmHistoryItem data type contains descriptive information about the history of a specific alarm. If you call DescribeAlarmHistory, Amazon CloudWatch returns this data type as part of the DescribeAlarmHistoryResult data type.

mkAlarmHistoryItem :: AlarmHistoryItem

Smart constructor for the minimum required fields to construct a valid AlarmHistoryItem data type.

AlarmHistoryItem is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

AlarmName :: Maybe Text

Timestamp :: Maybe ISO8601

HistoryItemType :: Maybe HistoryItemType

HistorySummary :: Maybe Text

HistoryData :: Maybe Text

ahiAlarmName :: Lens' AlarmHistoryItem (Maybe Text)

The descriptive name for the alarm.

ahiTimestamp :: Lens' AlarmHistoryItem (Maybe ISO8601)

The time stamp for the alarm history item.

ahiHistoryItemType :: Lens' AlarmHistoryItem (Maybe HistoryItemType)

The type of alarm history item.

ahiHistorySummary :: Lens' AlarmHistoryItem (Maybe Text)

A human-readable summary of the alarm history.

ahiHistoryData :: Lens' AlarmHistoryItem (Maybe Text)

Machine-readable data about the alarm in JSON format.

Datapoint

data Datapoint

The Datapoint data type encapsulates the statistical data that Amazon CloudWatch computes from metric data.

mkDatapoint :: Datapoint

Smart constructor for the minimum required fields to construct a valid Datapoint data type.

Datapoint is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Timestamp :: Maybe ISO8601

SampleCount :: Maybe Double

Average :: Maybe Double

Sum :: Maybe Double

Minimum :: Maybe Double

Maximum :: Maybe Double

Unit :: Maybe StandardUnit

drTimestamp :: Lens' Datapoint (Maybe ISO8601)

The time stamp used for the datapoint.

drSampleCount :: Lens' Datapoint (Maybe Double)

The number of metric values that contributed to the aggregate value of this datapoint.

drAverage :: Lens' Datapoint (Maybe Double)

The average of metric values that correspond to the datapoint.

drSum :: Lens' Datapoint (Maybe Double)

The sum of metric values used for the datapoint.

drMinimum :: Lens' Datapoint (Maybe Double)

The minimum metric value used for the datapoint.

drMaximum :: Lens' Datapoint (Maybe Double)

The maximum of the metric value used for the datapoint.

drUnit :: Lens' Datapoint (Maybe StandardUnit)

The standard unit used for the datapoint.

Dimension

data Dimension

The Dimension data type further expands on the identity of a metric using a Name, Value pair.

mkDimension

Arguments

:: Text
dName

-> Text
dValue

-> Dimension

Smart constructor for the minimum required fields to construct a valid Dimension data type to populate a request.

The fields accessible through corresponding lenses are:

Name :: Text

Value :: Text

dName :: Lens' Dimension Text

The name of the dimension.

dValue :: Lens' Dimension Text

The value representing the dimension measurement.

DimensionFilter

data DimensionFilter

The DimensionFilter data type is used to filter ListMetrics results.

mkDimensionFilter

Arguments

:: Text
dfName

-> DimensionFilter

Smart constructor for the minimum required fields to construct a valid DimensionFilter data type to populate a request.

The fields accessible through corresponding lenses are:

Name :: Text

Value :: Maybe Text

dfName :: Lens' DimensionFilter Text

The dimension name to be matched.

dfValue :: Lens' DimensionFilter (Maybe Text)

The value of the dimension to be matched. Specifying a Name without specifying a Value returns all values associated with that Name.

Metric

data Metric

The Metric data type contains information about a specific metric. If you call ListMetrics, Amazon CloudWatch returns information contained by this data type.

mkMetric :: Metric

Smart constructor for the minimum required fields to construct a valid Metric data type.

Metric is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Namespace :: Maybe Text

MetricName :: Maybe Text

Dimensions :: [Dimension]

mNamespace :: Lens' Metric (Maybe Text)

The namespace of the metric.

mMetricName :: Lens' Metric (Maybe Text)

The name of the metric.

mDimensions :: Lens' Metric [Dimension]

A list of dimensions associated with the metric.

MetricAlarm

data MetricAlarm

The MetricAlarm data type represents an alarm. You can use PutMetricAlarm to create or update an alarm.

mkMetricAlarm :: MetricAlarm

Smart constructor for the minimum required fields to construct a valid MetricAlarm data type.

MetricAlarm is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

AlarmName :: Maybe Text

AlarmArn :: Maybe Text

AlarmDescription :: Maybe Text

AlarmConfigurationUpdatedTimestamp :: Maybe ISO8601

ActionsEnabled :: Maybe Bool

OKActions :: [Text]

AlarmActions :: [Text]

InsufficientDataActions :: [Text]

StateValue :: Maybe StateValue

StateReason :: Maybe Text

StateReasonData :: Maybe Text

StateUpdatedTimestamp :: Maybe ISO8601

MetricName :: Maybe Text

Namespace :: Maybe Text

Statistic :: Maybe Statistic

Dimensions :: [Dimension]

Period :: Maybe Integer

Unit :: Maybe StandardUnit

EvaluationPeriods :: Maybe Integer

Threshold :: Maybe Double

ComparisonOperator :: Maybe ComparisonOperator

maAlarmName :: Lens' MetricAlarm (Maybe Text)

The name of the alarm.

maAlarmArn :: Lens' MetricAlarm (Maybe Text)

The Amazon Resource Name (ARN) of the alarm.

maAlarmDescription :: Lens' MetricAlarm (Maybe Text)

The description for the alarm.

maAlarmConfigurationUpdatedTimestamp :: Lens' MetricAlarm (Maybe ISO8601)

The time stamp of the last update to the alarm configuration.

maActionsEnabled :: Lens' MetricAlarm (Maybe Bool)

Indicates whether actions should be executed during any changes to the alarm's state.

maOKActions :: Lens' MetricAlarm [Text]

The list of actions to execute when this alarm transitions into an OK state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only actions supported are publishing to an Amazon SNS topic and triggering an Auto Scaling policy.

maAlarmActions :: Lens' MetricAlarm [Text]

The list of actions to execute when this alarm transitions into an ALARM state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only actions supported are publishing to an Amazon SNS topic and triggering an Auto Scaling policy.

maInsufficientDataActions :: Lens' MetricAlarm [Text]

The list of actions to execute when this alarm transitions into an INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only actions supported are publishing to an Amazon SNS topic or triggering an Auto Scaling policy.

maStateValue :: Lens' MetricAlarm (Maybe StateValue)

The state value for the alarm.

maStateReason :: Lens' MetricAlarm (Maybe Text)

A human-readable explanation for the alarm's state.

maStateReasonData :: Lens' MetricAlarm (Maybe Text)

An explanation for the alarm's state in machine-readable JSON format.

maStateUpdatedTimestamp :: Lens' MetricAlarm (Maybe ISO8601)

The time stamp of the last update to the alarm's state.

maMetricName :: Lens' MetricAlarm (Maybe Text)

The name of the alarm's metric.

maNamespace :: Lens' MetricAlarm (Maybe Text)

The namespace of alarm's associated metric.

maStatistic :: Lens' MetricAlarm (Maybe Statistic)

The statistic to apply to the alarm's associated metric.

maDimensions :: Lens' MetricAlarm [Dimension]

The list of dimensions associated with the alarm's associated metric.

maPeriod :: Lens' MetricAlarm (Maybe Integer)

The period in seconds over which the statistic is applied.

maUnit :: Lens' MetricAlarm (Maybe StandardUnit)

The unit of the alarm's associated metric.

maEvaluationPeriods :: Lens' MetricAlarm (Maybe Integer)

The number of periods over which data is compared to the specified threshold.

maThreshold :: Lens' MetricAlarm (Maybe Double)

The value against which the specified statistic is compared.

maComparisonOperator :: Lens' MetricAlarm (Maybe ComparisonOperator)

The arithmetic operation to use when comparing the specified Statistic and Threshold. The specified Statistic value is used as the first operand.

MetricDatum

data MetricDatum

The MetricDatum data type encapsulates the information sent with PutMetricData to either create a new metric or add new values to be aggregated into an existing metric.

mkMetricDatum

Arguments

:: Text
mdMetricName

-> MetricDatum

Smart constructor for the minimum required fields to construct a valid MetricDatum data type to populate a request.

The fields accessible through corresponding lenses are:

MetricName :: Text

Dimensions :: [Dimension]

Timestamp :: Maybe ISO8601

Value :: Maybe Double

StatisticValues :: Maybe StatisticSet

Unit :: Maybe StandardUnit

mdMetricName :: Lens' MetricDatum Text

The name of the metric.

mdDimensions :: Lens' MetricDatum [Dimension]

A list of dimensions associated with the metric.

mdTimestamp :: Lens' MetricDatum (Maybe ISO8601)

The time stamp used for the metric. If not specified, the default value is set to the time the metric data was received.

mdValue :: Lens' MetricDatum (Maybe Double)

The value for the metric. Although the Value parameter accepts numbers of type Double, Amazon CloudWatch truncates values with very large exponents. Values with base-10 exponents greater than 126 (1 x 10^126) are truncated. Likewise, values with base-10 exponents less than -130 (1 x 10^-130) are also truncated.

mdStatisticValues :: Lens' MetricDatum (Maybe StatisticSet)

A set of statistical values describing the metric.

mdUnit :: Lens' MetricDatum (Maybe StandardUnit)

The unit of the metric.

StatisticSet

data StatisticSet

A set of statistical values describing the metric.

mkStatisticSet

Arguments

:: Double
ssSampleCount

-> Double
ssSum

-> Double
ssMinimum

-> Double
ssMaximum

-> StatisticSet

Smart constructor for the minimum required fields to construct a valid StatisticSet data type to populate a request.

The fields accessible through corresponding lenses are:

SampleCount :: Double

Sum :: Double

Minimum :: Double

Maximum :: Double

ssSampleCount :: Lens' StatisticSet Double

The number of samples used for the statistic set.

ssSum :: Lens' StatisticSet Double

The sum of values for the sample set.

ssMinimum :: Lens' StatisticSet Double

The minimum value of the sample set.

ssMaximum :: Lens' StatisticSet Double

The maximum value of the sample set.