Network.AWS.CloudWatch.V2010_08_01.DescribeAlarms

Description

Retrieves alarms with the specified names. If no name is specified, all alarms for the user are returned. Alarms can be retrieved by using only a prefix for the alarm name, the alarm state, or a prefix for any action.

Documentation

describeAlarms :: DescribeAlarms

Minimum specification for a DescribeAlarms request.

data DescribeAlarms

Constructors

DescribeAlarms

Fields

_daiActionPrefix :: Maybe Text

The action name prefix.

_daiAlarmNamePrefix :: Maybe Text

The alarm name prefix. AlarmNames cannot be specified if this parameter is specified.

_daiAlarmNames :: [Text]

A list of alarm names to retrieve information for.

_daiMaxRecords :: Maybe Integer

The maximum number of alarm descriptions to retrieve.

_daiNextToken :: Maybe Text

The token returned by a previous call to indicate that there is more data available.

_daiStateValue :: Maybe StateValue

The state value to be used in matching alarms.

daiStateValue :: Lens' DescribeAlarms (Maybe StateValue)

daiNextToken :: Lens' DescribeAlarms (Maybe Text)

daiMaxRecords :: Lens' DescribeAlarms (Maybe Integer)

daiAlarmNames :: Lens' DescribeAlarms [Text]

daiAlarmNamePrefix :: Lens' DescribeAlarms (Maybe Text)

daiActionPrefix :: Lens' DescribeAlarms (Maybe Text)

data DescribeAlarmsResponse

Constructors

DescribeAlarmsResponse

Fields

_daoMetricAlarms :: [MetricAlarm]

A list of information for the specified alarms.

_daoNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

daoNextToken :: Lens' DescribeAlarmsResponse (Maybe Text)

daoMetricAlarms :: Lens' DescribeAlarmsResponse [MetricAlarm]