Network.AWS.CloudWatch.V2010_08_01.DescribeAlarmsForMetric

Description

Retrieves all alarms for a single metric. Specify a statistic, period, or unit to filter the set of alarms further.

Documentation

describeAlarmsForMetric

Arguments

:: Text
_dafmiMetricName

-> Text
_dafmiNamespace

-> DescribeAlarmsForMetric

Minimum specification for a DescribeAlarmsForMetric request.

data DescribeAlarmsForMetric

Constructors

DescribeAlarmsForMetric

Fields

_dafmiMetricName :: Text

The name of the metric.

_dafmiNamespace :: Text

The namespace of the metric.

_dafmiDimensions :: [Dimension]

The list of dimensions associated with the metric.

_dafmiPeriod :: Maybe Integer

The period in seconds over which the statistic is applied.

_dafmiUnit :: Maybe StandardUnit

The unit for the metric.

_dafmiStatistic :: Maybe Statistic

The statistic for the metric.

dafmiUnit :: Lens' DescribeAlarmsForMetric (Maybe StandardUnit)

dafmiStatistic :: Lens' DescribeAlarmsForMetric (Maybe Statistic)

dafmiPeriod :: Lens' DescribeAlarmsForMetric (Maybe Integer)

dafmiNamespace :: Lens' DescribeAlarmsForMetric Text

dafmiMetricName :: Lens' DescribeAlarmsForMetric Text

dafmiDimensions :: Lens' DescribeAlarmsForMetric [Dimension]

data DescribeAlarmsForMetricResponse

Constructors

DescribeAlarmsForMetricResponse

Fields

_dafmoMetricAlarms :: [MetricAlarm]

A list of information for each alarm with the specified metric.

dafmoMetricAlarms :: Iso' DescribeAlarmsForMetricResponse [MetricAlarm]