Network.AWS.CloudWatch.V2010_08_01.EnableAlarmActions

Description

Enables actions for the specified alarms.

Request

data EnableAlarmActions

Request constructor

mkEnableAlarmActions

Arguments

:: [Text]
eaaAlarmNames

-> EnableAlarmActions

Smart constructor for the minimum required parameters to construct a valid EnableAlarmActions request.

Request lenses

eaaAlarmNames :: Lens' EnableAlarmActions [Text]

The names of the alarms to enable actions for.

Response

data EnableAlarmActionsResponse

Response constructor

mkEnableAlarmActionsResponse :: EnableAlarmActionsResponse

Smart constructor for the minimum required parameters to construct a valid EnableAlarmActionsResponse response.

This constructor is provided for convenience and testing purposes.