Network.AWS.CloudWatch.V2010_08_01.PutMetricAlarm

Description

Creates or updates an alarm and associates it with the specified Amazon CloudWatch metric. Optionally, this operation can associate one or more Amazon Simple Notification Service resources with the alarm. When this operation creates an alarm, the alarm state is immediately set to INSUFFICIENT_DATA. The alarm is evaluated and its StateValue is set appropriately. Any actions associated with the StateValue is then executed. When updating an existing alarm, its StateValue is left unchanged.

Documentation

putMetricAlarm

Arguments

:: Text
_pmaiAlarmName

-> ComparisonOperator
_pmaiComparisonOperator

-> Integer
_pmaiEvaluationPeriods

-> Text
_pmaiMetricName

-> Text
_pmaiNamespace

-> Integer
_pmaiPeriod

-> Statistic
_pmaiStatistic

-> Double
_pmaiThreshold

-> PutMetricAlarm

Minimum specification for a PutMetricAlarm request.

data PutMetricAlarm

Constructors

PutMetricAlarm

Fields

_pmaiAlarmName :: Text

The descriptive name for the alarm. This name must be unique within the user's AWS account.

_pmaiComparisonOperator :: ComparisonOperator

The arithmetic operation to use when comparing the specified Statistic and Threshold. The specified Statistic value is used as the first operand.

_pmaiEvaluationPeriods :: Integer

The number of periods over which data is compared to the specified threshold.

_pmaiMetricName :: Text

The name for the alarm's associated metric.

_pmaiNamespace :: Text

The namespace for the alarm's associated metric.

_pmaiPeriod :: Integer

The period in seconds over which the specified statistic is applied.

_pmaiStatistic :: Statistic

The statistic to apply to the alarm's associated metric.

_pmaiThreshold :: Double

The value against which the specified statistic is compared.

_pmaiActionsEnabled :: Maybe Bool

Indicates whether or not actions should be executed during any changes to the alarm's state.

_pmaiAlarmDescription :: Maybe Text

The description for the alarm.

_pmaiDimensions :: [Dimension]

The dimensions for the alarm's associated metric.

_pmaiOKActions :: [Text]

The list of actions to execute when this alarm transitions into an OK state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only action supported is publishing to an Amazon SNS topic or an Amazon Auto Scaling policy.

_pmaiInsufficientDataActions :: [Text]

The list of actions to execute when this alarm transitions into an INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only action supported is publishing to an Amazon SNS topic or an Amazon Auto Scaling policy.

_pmaiAlarmActions :: [Text]

The list of actions to execute when this alarm transitions into an ALARM state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only action supported is publishing to an Amazon SNS topic or an Amazon Auto Scaling policy.

_pmaiUnit :: Maybe StandardUnit

The unit for the alarm's associated metric.

pmaiUnit :: Lens' PutMetricAlarm (Maybe StandardUnit)

pmaiThreshold :: Lens' PutMetricAlarm Double

pmaiStatistic :: Lens' PutMetricAlarm Statistic

pmaiPeriod :: Lens' PutMetricAlarm Integer

pmaiOKActions :: Lens' PutMetricAlarm [Text]

pmaiNamespace :: Lens' PutMetricAlarm Text

pmaiMetricName :: Lens' PutMetricAlarm Text

pmaiInsufficientDataActions :: Lens' PutMetricAlarm [Text]

pmaiEvaluationPeriods :: Lens' PutMetricAlarm Integer

pmaiDimensions :: Lens' PutMetricAlarm [Dimension]

pmaiComparisonOperator :: Lens' PutMetricAlarm ComparisonOperator

pmaiAlarmName :: Lens' PutMetricAlarm Text

pmaiAlarmDescription :: Lens' PutMetricAlarm (Maybe Text)

pmaiAlarmActions :: Lens' PutMetricAlarm [Text]

pmaiActionsEnabled :: Lens' PutMetricAlarm (Maybe Bool)

data PutMetricAlarmResponse

Constructors

PutMetricAlarmResponse