Network.AWS.CloudWatch.V2010_08_01.Types

Description

Amazon CloudWatch is a monitoring service for AWS cloud resources and the applications you run on AWS. You can use Amazon CloudWatch to collect and track metrics, collect and monitor log files, and set alarms. Amazon CloudWatch can monitor AWS resources such as Amazon EC2 instances, Amazon DynamoDB tables, and Amazon RDS DB instances, as well as custom metrics generated by your applications and services, and any log files your applications generate. You can use Amazon CloudWatch to gain system-wide visibility into resource utilization, application performance, and operational health. You can use these insights to react and keep your application running smoothly.

Documentation

data CloudWatch

Supported version (2010-08-01) of the Amazon CloudWatch service.

xmlOptions :: Tagged a XMLOptions

data ComparisonOperator

The arithmetic operation to use when comparing the specified Statistic and Threshold. The specified Statistic value is used as the first operand.

Constructors

ComparisonOperatorGreaterThanOrEqualToThreshold
GreaterThanOrEqualToThreshold

ComparisonOperatorGreaterThanThreshold
GreaterThanThreshold

ComparisonOperatorLessThanOrEqualToThreshold
LessThanOrEqualToThreshold

ComparisonOperatorLessThanThreshold
LessThanThreshold

data HistoryItemType

The type of alarm histories to retrieve.

Constructors

HistoryItemTypeAction
Action

HistoryItemTypeConfigurationUpdate
ConfigurationUpdate

HistoryItemTypeStateUpdate
StateUpdate

data StandardUnit

The unit of the metric.

Constructors

StandardUnitBits
Bits

StandardUnitBitsSecond
Bits/Second

StandardUnitBytes
Bytes

StandardUnitBytesSecond
Bytes/Second

StandardUnitCount
Count

StandardUnitCountSecond
Count/Second

StandardUnitGigabits
Gigabits

StandardUnitGigabitsSecond
Gigabits/Second

StandardUnitGigabytes
Gigabytes

StandardUnitGigabytesSecond
Gigabytes/Second

StandardUnitKilobits
Kilobits

StandardUnitKilobitsSecond
Kilobits/Second

StandardUnitKilobytes
Kilobytes

StandardUnitKilobytesSecond
Kilobytes/Second

StandardUnitMegabits
Megabits

StandardUnitMegabitsSecond
Megabits/Second

StandardUnitMegabytes
Megabytes

StandardUnitMegabytesSecond
Megabytes/Second

StandardUnitMicroseconds
Microseconds

StandardUnitMilliseconds
Milliseconds

StandardUnitNone
None

StandardUnitPercent
Percent

StandardUnitSeconds
Seconds

StandardUnitTerabits
Terabits

StandardUnitTerabitsSecond
Terabits/Second

StandardUnitTerabytes
Terabytes

StandardUnitTerabytesSecond
Terabytes/Second

data StateValue

The state value to be used in matching alarms.

Constructors

StateValueAlarm
ALARM

StateValueInsufficientData
INSUFFICIENT_DATA

StateValueOk
OK

data Statistic

The statistic to apply to the alarm's associated metric.

Constructors

StatisticAverage
Average

StatisticMaximum
Maximum

StatisticMinimum
Minimum

StatisticSampleCount
SampleCount

StatisticSum
Sum

data AlarmHistoryItem

The AlarmHistoryItem data type contains descriptive information about the history of a specific alarm. If you call DescribeAlarmHistory, Amazon CloudWatch returns this data type as part of the DescribeAlarmHistoryResult data type.

Constructors

AlarmHistoryItem

Fields

_ahjAlarmName :: Maybe Text

The descriptive name for the alarm.

_ahjHistoryItemType :: Maybe HistoryItemType

The type of alarm history item.

_ahjHistoryData :: Maybe Text

Machine-readable data about the alarm in JSON format.

_ahjHistorySummary :: Maybe Text

A human-readable summary of the alarm history.

_ahjTimestamp :: Maybe ISO8601

The time stamp for the alarm history item.

data Datapoint

The Datapoint data type encapsulates the statistical data that Amazon CloudWatch computes from metric data.

Constructors

Datapoint

Fields

_dtSampleCount :: Maybe Double

The number of metric values that contributed to the aggregate value of this datapoint.

_dtMaximum :: Maybe Double

The maximum of the metric value used for the datapoint.

_dtAverage :: Maybe Double

The average of metric values that correspond to the datapoint.

_dtMinimum :: Maybe Double

The minimum metric value used for the datapoint.

_dtSum :: Maybe Double

The sum of metric values used for the datapoint.

_dtUnit :: Maybe StandardUnit

The standard unit used for the datapoint.

_dtTimestamp :: Maybe ISO8601

The time stamp used for the datapoint.

data Dimension

The Dimension data type further expands on the identity of a metric using a Name, Value pair.

Constructors

Dimension

Fields

_eValue :: Text

The value representing the dimension measurement.

_eName :: Text

The name of the dimension.

data DimensionFilter

The DimensionFilter data type is used to filter ListMetrics results.

Constructors

DimensionFilter

Fields

_dgValue :: Maybe Text

The value of the dimension to be matched. Specifying a Name without specifying a Value returns all values associated with that Name.

_dgName :: Text

The dimension name to be matched.

data Metric

The Metric data type contains information about a specific metric. If you call ListMetrics, Amazon CloudWatch returns information contained by this data type.

Constructors

Metric

Fields

_rMetricName :: Maybe Text

The name of the metric.

_rNamespace :: Maybe Text

The namespace of the metric.

_rDimensions :: [Dimension]

A list of dimensions associated with the metric.

data MetricAlarm

The MetricAlarm data type represents an alarm. You can use PutMetricAlarm to create or update an alarm.

Constructors

MetricAlarm

Fields

_mbAlarmName :: Maybe Text

The name of the alarm.

_mbStateUpdatedTimestamp :: Maybe ISO8601

The time stamp of the last update to the alarm's state.

_mbPeriod :: Maybe Integer

The period in seconds over which the statistic is applied.

_mbAlarmDescription :: Maybe Text

The description for the alarm.

_mbEvaluationPeriods :: Maybe Integer

The number of periods over which data is compared to the specified threshold.

_mbMetricName :: Maybe Text

The name of the alarm's metric.

_mbNamespace :: Maybe Text

The namespace of alarm's associated metric.

_mbComparisonOperator :: Maybe ComparisonOperator

The arithmetic operation to use when comparing the specified Statistic and Threshold. The specified Statistic value is used as the first operand.

_mbOKActions :: [Text]

The list of actions to execute when this alarm transitions into an OK state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only actions supported are publishing to an Amazon SNS topic and triggering an Auto Scaling policy.

_mbStateValue :: Maybe StateValue

The state value for the alarm.

_mbThreshold :: Maybe Double

The value against which the specified statistic is compared.

_mbAlarmConfigurationUpdatedTimestamp :: Maybe ISO8601

The time stamp of the last update to the alarm configuration.

_mbActionsEnabled :: Maybe Bool

Indicates whether actions should be executed during any changes to the alarm's state.

_mbInsufficientDataActions :: [Text]

The list of actions to execute when this alarm transitions into an INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only actions supported are publishing to an Amazon SNS topic or triggering an Auto Scaling policy.

_mbStateReason :: Maybe Text

A human-readable explanation for the alarm's state.

_mbStateReasonData :: Maybe Text

An explanation for the alarm's state in machine-readable JSON format.

_mbDimensions :: [Dimension]

The list of dimensions associated with the alarm's associated metric.

_mbAlarmArn :: Maybe Text

The Amazon Resource Name (ARN) of the alarm.

_mbAlarmActions :: [Text]

The list of actions to execute when this alarm transitions into an ALARM state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only actions supported are publishing to an Amazon SNS topic and triggering an Auto Scaling policy.

_mbUnit :: Maybe StandardUnit

The unit of the alarm's associated metric.

_mbStatistic :: Maybe Statistic

The statistic to apply to the alarm's associated metric.

data MetricDatum

The MetricDatum data type encapsulates the information sent with PutMetricData to either create a new metric or add new values to be aggregated into an existing metric.

Constructors

MetricDatum

Fields

_meMetricName :: Text

The name of the metric.

_meValue :: Maybe Double

The value for the metric. Although the Value parameter accepts numbers of type Double, Amazon CloudWatch truncates values with very large exponents. Values with base-10 exponents greater than 126 (1 x 10^126) are truncated. Likewise, values with base-10 exponents less than -130 (1 x 10^-130) are also truncated.

_meDimensions :: [Dimension]

A list of dimensions associated with the metric.

_meUnit :: Maybe StandardUnit

The unit of the metric.

_meTimestamp :: Maybe ISO8601

The time stamp used for the metric. If not specified, the default value is set to the time the metric data was received.

_meStatisticValues :: Maybe StatisticSet

A set of statistical values describing the metric.

data StatisticSet

A set of statistical values describing the metric.

Constructors

StatisticSet

Fields

_ssSampleCount :: Double

The number of samples used for the statistic set.

_ssMaximum :: Double

The maximum value of the sample set.

_ssMinimum :: Double

The minimum value of the sample set.

_ssSum :: Double

The sum of values for the sample set.

ahjTimestamp :: Lens' AlarmHistoryItem (Maybe ISO8601)

ahjHistorySummary :: Lens' AlarmHistoryItem (Maybe Text)

ahjHistoryItemType :: Lens' AlarmHistoryItem (Maybe HistoryItemType)

ahjHistoryData :: Lens' AlarmHistoryItem (Maybe Text)

ahjAlarmName :: Lens' AlarmHistoryItem (Maybe Text)

dtUnit :: Lens' Datapoint (Maybe StandardUnit)

dtTimestamp :: Lens' Datapoint (Maybe ISO8601)

dtSum :: Lens' Datapoint (Maybe Double)

dtSampleCount :: Lens' Datapoint (Maybe Double)

dtMinimum :: Lens' Datapoint (Maybe Double)

dtMaximum :: Lens' Datapoint (Maybe Double)

dtAverage :: Lens' Datapoint (Maybe Double)

eValue :: Lens' Dimension Text

eName :: Lens' Dimension Text

dgValue :: Lens' DimensionFilter (Maybe Text)

dgName :: Lens' DimensionFilter Text

rNamespace :: Lens' Metric (Maybe Text)

rMetricName :: Lens' Metric (Maybe Text)

rDimensions :: Lens' Metric [Dimension]

mbUnit :: Lens' MetricAlarm (Maybe StandardUnit)

mbThreshold :: Lens' MetricAlarm (Maybe Double)

mbStatistic :: Lens' MetricAlarm (Maybe Statistic)

mbStateValue :: Lens' MetricAlarm (Maybe StateValue)

mbStateUpdatedTimestamp :: Lens' MetricAlarm (Maybe ISO8601)

mbStateReasonData :: Lens' MetricAlarm (Maybe Text)

mbStateReason :: Lens' MetricAlarm (Maybe Text)

mbPeriod :: Lens' MetricAlarm (Maybe Integer)

mbOKActions :: Lens' MetricAlarm [Text]

mbNamespace :: Lens' MetricAlarm (Maybe Text)

mbMetricName :: Lens' MetricAlarm (Maybe Text)

mbInsufficientDataActions :: Lens' MetricAlarm [Text]

mbEvaluationPeriods :: Lens' MetricAlarm (Maybe Integer)

mbDimensions :: Lens' MetricAlarm [Dimension]

mbComparisonOperator :: Lens' MetricAlarm (Maybe ComparisonOperator)

mbAlarmName :: Lens' MetricAlarm (Maybe Text)

mbAlarmDescription :: Lens' MetricAlarm (Maybe Text)

mbAlarmConfigurationUpdatedTimestamp :: Lens' MetricAlarm (Maybe ISO8601)

mbAlarmArn :: Lens' MetricAlarm (Maybe Text)

mbAlarmActions :: Lens' MetricAlarm [Text]

mbActionsEnabled :: Lens' MetricAlarm (Maybe Bool)

meValue :: Lens' MetricDatum (Maybe Double)

meUnit :: Lens' MetricDatum (Maybe StandardUnit)

meTimestamp :: Lens' MetricDatum (Maybe ISO8601)

meStatisticValues :: Lens' MetricDatum (Maybe StatisticSet)

meMetricName :: Lens' MetricDatum Text

meDimensions :: Lens' MetricDatum [Dimension]

ssSum :: Lens' StatisticSet Double

ssSampleCount :: Lens' StatisticSet Double

ssMinimum :: Lens' StatisticSet Double

ssMaximum :: Lens' StatisticSet Double