Network.AWS.CloudWatch.DescribeAlarmHistory

Description

Retrieves the history for the specified alarm. You can filter the results by date range or item type. If an alarm name is not specified, the histories for all alarms are returned.

CloudWatch retains the history of an alarm even if you delete the alarm.

This operation returns paginated results.

Creating a Request

describeAlarmHistory :: DescribeAlarmHistory

Creates a value of DescribeAlarmHistory with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dahAlarmName - The name of the alarm.

dahHistoryItemType - The type of alarm histories to retrieve.

dahEndDate - The ending date to retrieve alarm history.

dahStartDate - The starting date to retrieve alarm history.

dahNextToken - The token returned by a previous call to indicate that there is more data available.

dahMaxRecords - The maximum number of alarm history records to retrieve.

data DescribeAlarmHistory

See: describeAlarmHistory smart constructor.

Request Lenses

dahAlarmName :: Lens' DescribeAlarmHistory (Maybe Text)

The name of the alarm.

dahHistoryItemType :: Lens' DescribeAlarmHistory (Maybe HistoryItemType)

The type of alarm histories to retrieve.

dahEndDate :: Lens' DescribeAlarmHistory (Maybe UTCTime)

The ending date to retrieve alarm history.

dahStartDate :: Lens' DescribeAlarmHistory (Maybe UTCTime)

The starting date to retrieve alarm history.

dahNextToken :: Lens' DescribeAlarmHistory (Maybe Text)

The token returned by a previous call to indicate that there is more data available.

dahMaxRecords :: Lens' DescribeAlarmHistory (Maybe Natural)

The maximum number of alarm history records to retrieve.

Destructuring the Response

describeAlarmHistoryResponse

Arguments

:: Int
dahrsResponseStatus

-> DescribeAlarmHistoryResponse

Creates a value of DescribeAlarmHistoryResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dahrsAlarmHistoryItems - The alarm histories, in JSON format.

dahrsNextToken - The token that marks the start of the next batch of returned results.

dahrsResponseStatus - -- | The response status code.

data DescribeAlarmHistoryResponse

See: describeAlarmHistoryResponse smart constructor.

Response Lenses

dahrsAlarmHistoryItems :: Lens' DescribeAlarmHistoryResponse [AlarmHistoryItem]

The alarm histories, in JSON format.

dahrsNextToken :: Lens' DescribeAlarmHistoryResponse (Maybe Text)

The token that marks the start of the next batch of returned results.

dahrsResponseStatus :: Lens' DescribeAlarmHistoryResponse Int

- | The response status code.