Network.AWS.CloudWatch.DescribeAlarmHistory

Description

Retrieves history for the specified alarm. Filter alarms by date range or item type. If an alarm name is not specified, Amazon CloudWatch returns histories for all of the owner's alarms.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_DescribeAlarmHistory.html

Request

data DescribeAlarmHistory

Request constructor

describeAlarmHistory :: DescribeAlarmHistory

DescribeAlarmHistory constructor.

The fields accessible through corresponding lenses are:

dahAlarmName :: Maybe Text

dahEndDate :: Maybe UTCTime

dahHistoryItemType :: Maybe HistoryItemType

dahMaxRecords :: Maybe Natural

dahNextToken :: Maybe Text

dahStartDate :: Maybe UTCTime

Request lenses

dahAlarmName :: Lens' DescribeAlarmHistory (Maybe Text)

The name of the alarm.

dahEndDate :: Lens' DescribeAlarmHistory (Maybe UTCTime)

The ending date to retrieve alarm history.

dahHistoryItemType :: Lens' DescribeAlarmHistory (Maybe HistoryItemType)

The type of alarm histories to retrieve.

dahMaxRecords :: Lens' DescribeAlarmHistory (Maybe Natural)

The maximum number of alarm history records to retrieve.

dahNextToken :: Lens' DescribeAlarmHistory (Maybe Text)

The token returned by a previous call to indicate that there is more data available.

dahStartDate :: Lens' DescribeAlarmHistory (Maybe UTCTime)

The starting date to retrieve alarm history.

Response

data DescribeAlarmHistoryResponse

Response constructor

describeAlarmHistoryResponse :: DescribeAlarmHistoryResponse

DescribeAlarmHistoryResponse constructor.

The fields accessible through corresponding lenses are:

dahrAlarmHistoryItems :: [AlarmHistoryItem]

dahrNextToken :: Maybe Text

Response lenses

dahrAlarmHistoryItems :: Lens' DescribeAlarmHistoryResponse [AlarmHistoryItem]

A list of alarm histories in JSON format.

dahrNextToken :: Lens' DescribeAlarmHistoryResponse (Maybe Text)

A string that marks the start of the next batch of returned results.