Network.AWS.CloudWatch.DescribeAlarms

Description

Retrieves the specified alarms. If no alarms are specified, all alarms are returned. Alarms can be retrieved by using only a prefix for the alarm name, the alarm state, or a prefix for any action.

This operation returns paginated results.

Creating a Request

describeAlarms :: DescribeAlarms

Creates a value of DescribeAlarms with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

daAlarmNamePrefix - The alarm name prefix. If this parameter is specified, you cannot specify AlarmNames .

daActionPrefix - The action name prefix.

daNextToken - The token returned by a previous call to indicate that there is more data available.

daStateValue - The state value to be used in matching alarms.

daAlarmNames - The names of the alarms.

daMaxRecords - The maximum number of alarm descriptions to retrieve.

data DescribeAlarms

See: describeAlarms smart constructor.

Request Lenses

daAlarmNamePrefix :: Lens' DescribeAlarms (Maybe Text)

The alarm name prefix. If this parameter is specified, you cannot specify AlarmNames .

daActionPrefix :: Lens' DescribeAlarms (Maybe Text)

The action name prefix.

daNextToken :: Lens' DescribeAlarms (Maybe Text)

The token returned by a previous call to indicate that there is more data available.

daStateValue :: Lens' DescribeAlarms (Maybe StateValue)

The state value to be used in matching alarms.

daAlarmNames :: Lens' DescribeAlarms [Text]

The names of the alarms.

daMaxRecords :: Lens' DescribeAlarms (Maybe Natural)

The maximum number of alarm descriptions to retrieve.

Destructuring the Response

describeAlarmsResponse

Arguments

:: Int
darsResponseStatus

-> DescribeAlarmsResponse

Creates a value of DescribeAlarmsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

darsMetricAlarms - The information for the specified alarms.

darsNextToken - The token that marks the start of the next batch of returned results.

darsResponseStatus - -- | The response status code.

data DescribeAlarmsResponse

See: describeAlarmsResponse smart constructor.

Response Lenses

darsMetricAlarms :: Lens' DescribeAlarmsResponse [MetricAlarm]

The information for the specified alarms.

darsNextToken :: Lens' DescribeAlarmsResponse (Maybe Text)

The token that marks the start of the next batch of returned results.

darsResponseStatus :: Lens' DescribeAlarmsResponse Int

- | The response status code.