Network.AWS.CloudWatch.DescribeAlarms

Description

Retrieves alarms with the specified names. If no name is specified, all alarms for the user are returned. Alarms can be retrieved by using only a prefix for the alarm name, the alarm state, or a prefix for any action.

http://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_DescribeAlarms.html

Request

data DescribeAlarms

Request constructor

describeAlarms :: DescribeAlarms

DescribeAlarms constructor.

The fields accessible through corresponding lenses are:

daActionPrefix :: Maybe Text

daAlarmNamePrefix :: Maybe Text

daAlarmNames :: [Text]

daMaxRecords :: Maybe Natural

daNextToken :: Maybe Text

daStateValue :: Maybe StateValue

Request lenses

daActionPrefix :: Lens' DescribeAlarms (Maybe Text)

The action name prefix.

daAlarmNamePrefix :: Lens' DescribeAlarms (Maybe Text)

The alarm name prefix. AlarmNames cannot be specified if this parameter is specified.

daAlarmNames :: Lens' DescribeAlarms [Text]

A list of alarm names to retrieve information for.

daMaxRecords :: Lens' DescribeAlarms (Maybe Natural)

The maximum number of alarm descriptions to retrieve.

daNextToken :: Lens' DescribeAlarms (Maybe Text)

The token returned by a previous call to indicate that there is more data available.

daStateValue :: Lens' DescribeAlarms (Maybe StateValue)

The state value to be used in matching alarms.

Response

data DescribeAlarmsResponse

Response constructor

describeAlarmsResponse :: DescribeAlarmsResponse

DescribeAlarmsResponse constructor.

The fields accessible through corresponding lenses are:

darMetricAlarms :: [MetricAlarm]

darNextToken :: Maybe Text

Response lenses

darMetricAlarms :: Lens' DescribeAlarmsResponse [MetricAlarm]

A list of information for the specified alarms.

darNextToken :: Lens' DescribeAlarmsResponse (Maybe Text)

A string that marks the start of the next batch of returned results.