Network.AWS.CloudWatch.DescribeAlarmsForMetric

Description

Retrieves the alarms for the specified metric. To filter the results, specify a statistic, period, or unit.

Creating a Request

describeAlarmsForMetric

Arguments

:: Text
dafmMetricName

-> Text
dafmNamespace

-> DescribeAlarmsForMetric

Creates a value of DescribeAlarmsForMetric with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dafmPeriod - The period, in seconds, over which the statistic is applied.

dafmDimensions - The dimensions associated with the metric. If the metric has any associated dimensions, you must specify them in order for the call to succeed.

dafmUnit - The unit for the metric.

dafmStatistic - The statistic for the metric, other than percentiles. For percentile statistics, use ExtendedStatistics .

dafmExtendedStatistic - The percentile statistic for the metric. Specify a value between p0.0 and p100.

dafmMetricName - The name of the metric.

dafmNamespace - The namespace of the metric.

data DescribeAlarmsForMetric

See: describeAlarmsForMetric smart constructor.

Request Lenses

dafmPeriod :: Lens' DescribeAlarmsForMetric (Maybe Natural)

The period, in seconds, over which the statistic is applied.

dafmDimensions :: Lens' DescribeAlarmsForMetric [Dimension]

The dimensions associated with the metric. If the metric has any associated dimensions, you must specify them in order for the call to succeed.

dafmUnit :: Lens' DescribeAlarmsForMetric (Maybe StandardUnit)

The unit for the metric.

dafmStatistic :: Lens' DescribeAlarmsForMetric (Maybe Statistic)

The statistic for the metric, other than percentiles. For percentile statistics, use ExtendedStatistics .

dafmExtendedStatistic :: Lens' DescribeAlarmsForMetric (Maybe Text)

The percentile statistic for the metric. Specify a value between p0.0 and p100.

dafmMetricName :: Lens' DescribeAlarmsForMetric Text

The name of the metric.

dafmNamespace :: Lens' DescribeAlarmsForMetric Text

The namespace of the metric.

Destructuring the Response

describeAlarmsForMetricResponse

Arguments

:: Int
dafmrsResponseStatus

-> DescribeAlarmsForMetricResponse

Creates a value of DescribeAlarmsForMetricResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dafmrsMetricAlarms - The information for each alarm with the specified metric.

dafmrsResponseStatus - -- | The response status code.

data DescribeAlarmsForMetricResponse

See: describeAlarmsForMetricResponse smart constructor.

Response Lenses

dafmrsMetricAlarms :: Lens' DescribeAlarmsForMetricResponse [MetricAlarm]

The information for each alarm with the specified metric.

dafmrsResponseStatus :: Lens' DescribeAlarmsForMetricResponse Int

- | The response status code.