Network.AWS.CloudWatch.ListMetrics

Description

List the specified metrics. You can use the returned metrics with GetMetricStatistics to obtain statistical data.

Up to 500 results are returned for any one call. To retrieve additional results, use the returned token with subsequent calls.

After you create a metric, allow up to fifteen minutes before the metric appears. Statistics about the metric, however, are available sooner using GetMetricStatistics .

This operation returns paginated results.

Creating a Request

listMetrics :: ListMetrics

Creates a value of ListMetrics with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lmMetricName - The name of the metric to filter against.

lmNamespace - The namespace to filter against.

lmNextToken - The token returned by a previous call to indicate that there is more data available.

lmDimensions - The dimensions to filter against.

data ListMetrics

See: listMetrics smart constructor.

Request Lenses

lmMetricName :: Lens' ListMetrics (Maybe Text)

The name of the metric to filter against.

lmNamespace :: Lens' ListMetrics (Maybe Text)

The namespace to filter against.

lmNextToken :: Lens' ListMetrics (Maybe Text)

The token returned by a previous call to indicate that there is more data available.

lmDimensions :: Lens' ListMetrics [DimensionFilter]

The dimensions to filter against.

Destructuring the Response

listMetricsResponse

Arguments

:: Int
lmrsResponseStatus

-> ListMetricsResponse

Creates a value of ListMetricsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lmrsMetrics - The metrics.

lmrsNextToken - The token that marks the start of the next batch of returned results.

lmrsResponseStatus - -- | The response status code.

data ListMetricsResponse

See: listMetricsResponse smart constructor.

Response Lenses

lmrsMetrics :: Lens' ListMetricsResponse [Metric]

The metrics.

lmrsNextToken :: Lens' ListMetricsResponse (Maybe Text)

The token that marks the start of the next batch of returned results.

lmrsResponseStatus :: Lens' ListMetricsResponse Int

- | The response status code.