Network.AWS.CloudWatch.PutMetricAlarm

Description

Creates or updates an alarm and associates it with the specified metric. Optionally, this operation can associate one or more Amazon SNS resources with the alarm.

When this operation creates an alarm, the alarm state is immediately set to INSUFFICIENT_DATA . The alarm is evaluated and its state is set appropriately. Any actions associated with the state are then executed.

When you update an existing alarm, its state is left unchanged, but the update completely overwrites the previous configuration of the alarm.

If you are an IAM user, you must have Amazon EC2 permissions for some operations:

ec2:DescribeInstanceStatus and ec2:DescribeInstances for all alarms on EC2 instance status metrics

ec2:StopInstances for alarms with stop actions

ec2:TerminateInstances for alarms with terminate actions

ec2:DescribeInstanceRecoveryAttribute and ec2:RecoverInstances for alarms with recover actions

If you have read/write permissions for Amazon CloudWatch but not for Amazon EC2, you can still create an alarm, but the stop or terminate actions are not performed. However, if you are later granted the required permissions, the alarm actions that you created earlier are performed.

If you are using an IAM role (for example, an EC2 instance profile), you cannot stop or terminate the instance using alarm actions. However, you can still see the alarm state and perform any other actions such as Amazon SNS notifications or Auto Scaling policies.

If you are using temporary security credentials granted using AWS STS, you cannot stop or terminate an EC2 instance using alarm actions.

You must create at least one stop, terminate, or reboot alarm using either the Amazon EC2 or CloudWatch consoles to create the EC2ActionsAccess IAM role. After this IAM role is created, you can create stop, terminate, or reboot alarms using a command-line interface or API.

Creating a Request

putMetricAlarm

Arguments

:: Text
pmaAlarmName

-> Text
pmaMetricName

-> Text
pmaNamespace

-> Natural
pmaPeriod

-> Natural
pmaEvaluationPeriods

-> Double
pmaThreshold

-> ComparisonOperator
pmaComparisonOperator

-> PutMetricAlarm

Creates a value of PutMetricAlarm with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pmaTreatMissingData - Sets how this alarm is to handle missing data points. If TreatMissingData is omitted, the default behavior of missing is used. For more information, see Configuring How CloudWatch Alarms Treats Missing Data . Valid Values: breaching | notBreaching | ignore | missing

pmaAlarmDescription - The description for the alarm.

pmaOKActions - The actions to execute when this alarm transitions to an OK state from any other state. Each action is specified as an Amazon Resource Name (ARN). Valid Values: arn:aws:automate:region :ec2:stop | arn:aws:automate:region :ec2:terminate | arn:aws:automate:region :ec2:recover Valid Values (for use with IAM roles): arn:aws:swf:us-east-1:{customer-account }:actionactionsAWS_EC2.InstanceId.Stop1.0 | arn:aws:swf:us-east-1:{customer-account }:actionactionsAWS_EC2.InstanceId.Terminate1.0 | arn:aws:swf:us-east-1:{customer-account }:actionactionsAWS_EC2.InstanceId.Reboot/1.0

pmaEvaluateLowSampleCountPercentile - Used only for alarms based on percentiles. If you specify ignore , the alarm state does not change during periods with too few data points to be statistically significant. If you specify evaluate or omit this parameter, the alarm is always evaluated and possibly changes state no matter how many data points are available. For more information, see Percentile-Based CloudWatch Alarms and Low Data Samples . Valid Values: evaluate | ignore

pmaActionsEnabled - Indicates whether actions should be executed during any changes to the alarm state.

pmaInsufficientDataActions - The actions to execute when this alarm transitions to the INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Name (ARN). Valid Values: arn:aws:automate:region :ec2:stop | arn:aws:automate:region :ec2:terminate | arn:aws:automate:region :ec2:recover Valid Values (for use with IAM roles): arn:aws:swf:us-east-1:{customer-account }:actionactionsAWS_EC2.InstanceId.Stop1.0 | arn:aws:swf:us-east-1:{customer-account }:actionactionsAWS_EC2.InstanceId.Terminate1.0 | arn:aws:swf:us-east-1:{customer-account }:actionactionsAWS_EC2.InstanceId.Reboot/1.0

pmaDimensions - The dimensions for the metric associated with the alarm.

pmaAlarmActions - The actions to execute when this alarm transitions to the ALARM state from any other state. Each action is specified as an Amazon Resource Name (ARN). Valid Values: arn:aws:automate:region :ec2:stop | arn:aws:automate:region :ec2:terminate | arn:aws:automate:region :ec2:recover Valid Values (for use with IAM roles): arn:aws:swf:us-east-1:{customer-account }:actionactionsAWS_EC2.InstanceId.Stop1.0 | arn:aws:swf:us-east-1:{customer-account }:actionactionsAWS_EC2.InstanceId.Terminate1.0 | arn:aws:swf:us-east-1:{customer-account }:actionactionsAWS_EC2.InstanceId.Reboot/1.0

pmaUnit - The unit of measure for the statistic. For example, the units for the Amazon EC2 NetworkIn metric are Bytes because NetworkIn tracks the number of bytes that an instance receives on all network interfaces. You can also specify a unit when you create a custom metric. Units help provide conceptual meaning to your data. Metric data points that specify a unit of measure, such as Percent, are aggregated separately. If you specify a unit, you must use a unit that is appropriate for the metric. Otherwise, the CloudWatch alarm can get stuck in the INSUFFICIENT DATA state.

pmaStatistic - The statistic for the metric associated with the alarm, other than percentile. For percentile statistics, use ExtendedStatistic .

pmaExtendedStatistic - The percentile statistic for the metric associated with the alarm. Specify a value between p0.0 and p100.

pmaAlarmName - The name for the alarm. This name must be unique within the AWS account.

pmaMetricName - The name for the metric associated with the alarm.

pmaNamespace - The namespace for the metric associated with the alarm.

pmaPeriod - The period, in seconds, over which the specified statistic is applied. Valid values are 10, 30, and any multiple of 60. Be sure to specify 10 or 30 only for metrics that are stored by a PutMetricData call with a StorageResolution of 1. If you specify a Period of 10 or 30 for a metric that does not have sub-minute resolution, the alarm still attempts to gather data at the period rate that you specify. In this case, it does not receive data for the attempts that do not correspond to a one-minute data resolution, and the alarm may often lapse into INSUFFICENT_DATA status. Specifying 10 or 30 also sets this alarm as a high-resolution alarm, which has a higher charge than other alarms. For more information about pricing, see Amazon CloudWatch Pricing . An alarm's total current evaluation period can be no longer than one day, so Period multiplied by EvaluationPeriods cannot be more than 86,400 seconds.

pmaEvaluationPeriods - The number of periods over which data is compared to the specified threshold. An alarm's total current evaluation period can be no longer than one day, so this number multiplied by Period cannot be more than 86,400 seconds.

pmaThreshold - The value against which the specified statistic is compared.

pmaComparisonOperator - The arithmetic operation to use when comparing the specified statistic and threshold. The specified statistic value is used as the first operand.

data PutMetricAlarm

See: putMetricAlarm smart constructor.

Request Lenses

pmaTreatMissingData :: Lens' PutMetricAlarm (Maybe Text)

Sets how this alarm is to handle missing data points. If TreatMissingData is omitted, the default behavior of missing is used. For more information, see Configuring How CloudWatch Alarms Treats Missing Data . Valid Values: breaching | notBreaching | ignore | missing

pmaAlarmDescription :: Lens' PutMetricAlarm (Maybe Text)

The description for the alarm.

pmaOKActions :: Lens' PutMetricAlarm [Text]

The actions to execute when this alarm transitions to an OK state from any other state. Each action is specified as an Amazon Resource Name (ARN). Valid Values: arn:aws:automate:region :ec2:stop | arn:aws:automate:region :ec2:terminate | arn:aws:automate:region :ec2:recover Valid Values (for use with IAM roles): arn:aws:swf:us-east-1:{customer-account }:actionactionsAWS_EC2.InstanceId.Stop1.0 | arn:aws:swf:us-east-1:{customer-account }:actionactionsAWS_EC2.InstanceId.Terminate1.0 | arn:aws:swf:us-east-1:{customer-account }:actionactionsAWS_EC2.InstanceId.Reboot/1.0

pmaEvaluateLowSampleCountPercentile :: Lens' PutMetricAlarm (Maybe Text)

Used only for alarms based on percentiles. If you specify ignore , the alarm state does not change during periods with too few data points to be statistically significant. If you specify evaluate or omit this parameter, the alarm is always evaluated and possibly changes state no matter how many data points are available. For more information, see Percentile-Based CloudWatch Alarms and Low Data Samples . Valid Values: evaluate | ignore

pmaActionsEnabled :: Lens' PutMetricAlarm (Maybe Bool)

Indicates whether actions should be executed during any changes to the alarm state.

pmaInsufficientDataActions :: Lens' PutMetricAlarm [Text]

The actions to execute when this alarm transitions to the INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Name (ARN). Valid Values: arn:aws:automate:region :ec2:stop | arn:aws:automate:region :ec2:terminate | arn:aws:automate:region :ec2:recover Valid Values (for use with IAM roles): arn:aws:swf:us-east-1:{customer-account }:actionactionsAWS_EC2.InstanceId.Stop1.0 | arn:aws:swf:us-east-1:{customer-account }:actionactionsAWS_EC2.InstanceId.Terminate1.0 | arn:aws:swf:us-east-1:{customer-account }:actionactionsAWS_EC2.InstanceId.Reboot/1.0

pmaDimensions :: Lens' PutMetricAlarm [Dimension]

The dimensions for the metric associated with the alarm.

pmaAlarmActions :: Lens' PutMetricAlarm [Text]

The actions to execute when this alarm transitions to the ALARM state from any other state. Each action is specified as an Amazon Resource Name (ARN). Valid Values: arn:aws:automate:region :ec2:stop | arn:aws:automate:region :ec2:terminate | arn:aws:automate:region :ec2:recover Valid Values (for use with IAM roles): arn:aws:swf:us-east-1:{customer-account }:actionactionsAWS_EC2.InstanceId.Stop1.0 | arn:aws:swf:us-east-1:{customer-account }:actionactionsAWS_EC2.InstanceId.Terminate1.0 | arn:aws:swf:us-east-1:{customer-account }:actionactionsAWS_EC2.InstanceId.Reboot/1.0

pmaUnit :: Lens' PutMetricAlarm (Maybe StandardUnit)

The unit of measure for the statistic. For example, the units for the Amazon EC2 NetworkIn metric are Bytes because NetworkIn tracks the number of bytes that an instance receives on all network interfaces. You can also specify a unit when you create a custom metric. Units help provide conceptual meaning to your data. Metric data points that specify a unit of measure, such as Percent, are aggregated separately. If you specify a unit, you must use a unit that is appropriate for the metric. Otherwise, the CloudWatch alarm can get stuck in the INSUFFICIENT DATA state.

pmaStatistic :: Lens' PutMetricAlarm (Maybe Statistic)

The statistic for the metric associated with the alarm, other than percentile. For percentile statistics, use ExtendedStatistic .

pmaExtendedStatistic :: Lens' PutMetricAlarm (Maybe Text)

The percentile statistic for the metric associated with the alarm. Specify a value between p0.0 and p100.

pmaAlarmName :: Lens' PutMetricAlarm Text

The name for the alarm. This name must be unique within the AWS account.

pmaMetricName :: Lens' PutMetricAlarm Text

The name for the metric associated with the alarm.

pmaNamespace :: Lens' PutMetricAlarm Text

The namespace for the metric associated with the alarm.

pmaPeriod :: Lens' PutMetricAlarm Natural

The period, in seconds, over which the specified statistic is applied. Valid values are 10, 30, and any multiple of 60. Be sure to specify 10 or 30 only for metrics that are stored by a PutMetricData call with a StorageResolution of 1. If you specify a Period of 10 or 30 for a metric that does not have sub-minute resolution, the alarm still attempts to gather data at the period rate that you specify. In this case, it does not receive data for the attempts that do not correspond to a one-minute data resolution, and the alarm may often lapse into INSUFFICENT_DATA status. Specifying 10 or 30 also sets this alarm as a high-resolution alarm, which has a higher charge than other alarms. For more information about pricing, see Amazon CloudWatch Pricing . An alarm's total current evaluation period can be no longer than one day, so Period multiplied by EvaluationPeriods cannot be more than 86,400 seconds.

pmaEvaluationPeriods :: Lens' PutMetricAlarm Natural

The number of periods over which data is compared to the specified threshold. An alarm's total current evaluation period can be no longer than one day, so this number multiplied by Period cannot be more than 86,400 seconds.

pmaThreshold :: Lens' PutMetricAlarm Double

The value against which the specified statistic is compared.

pmaComparisonOperator :: Lens' PutMetricAlarm ComparisonOperator

The arithmetic operation to use when comparing the specified statistic and threshold. The specified statistic value is used as the first operand.

Destructuring the Response

putMetricAlarmResponse :: PutMetricAlarmResponse

Creates a value of PutMetricAlarmResponse with the minimum fields required to make a request.

data PutMetricAlarmResponse

See: putMetricAlarmResponse smart constructor.