Network.AWS.CloudWatch.SetAlarmState

Description

Temporarily sets the state of an alarm for testing purposes. When the updated state differs from the previous value, the action configured for the appropriate state is invoked. For example, if your alarm is configured to send an Amazon SNS message when an alarm is triggered, temporarily changing the alarm state to ALARM sends an SNS message. The alarm returns to its actual state (often within seconds). Because the alarm state change happens quickly, it is typically only visible in the alarm's History tab in the Amazon CloudWatch console or through DescribeAlarmHistory .

Creating a Request

setAlarmState

Arguments

:: Text
sasAlarmName

-> StateValue
sasStateValue

-> Text
sasStateReason

-> SetAlarmState

Creates a value of SetAlarmState with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sasStateReasonData - The reason that this alarm is set to this specific state, in JSON format.

sasAlarmName - The name for the alarm. This name must be unique within the AWS account. The maximum length is 255 characters.

sasStateValue - The value of the state.

sasStateReason - The reason that this alarm is set to this specific state, in text format.

data SetAlarmState

See: setAlarmState smart constructor.

Request Lenses

sasStateReasonData :: Lens' SetAlarmState (Maybe Text)

The reason that this alarm is set to this specific state, in JSON format.

sasAlarmName :: Lens' SetAlarmState Text

The name for the alarm. This name must be unique within the AWS account. The maximum length is 255 characters.

sasStateValue :: Lens' SetAlarmState StateValue

The value of the state.

sasStateReason :: Lens' SetAlarmState Text

The reason that this alarm is set to this specific state, in text format.

Destructuring the Response

setAlarmStateResponse :: SetAlarmStateResponse

Creates a value of SetAlarmStateResponse with the minimum fields required to make a request.

data SetAlarmStateResponse

See: setAlarmStateResponse smart constructor.