Network.AWS.CloudWatch.Types

Description

Service Configuration

cloudWatch :: Service

API version 2010-08-01 of the Amazon CloudWatch SDK configuration.

Errors

_LimitExceededFault :: AsError a => Getting (First ServiceError) a ServiceError

The quota for alarms for this customer has already been reached.

_DashboardNotFoundError :: AsError a => Getting (First ServiceError) a ServiceError

The specified dashboard does not exist.

_InvalidNextToken :: AsError a => Getting (First ServiceError) a ServiceError

The next token specified is invalid.

_InternalServiceFault :: AsError a => Getting (First ServiceError) a ServiceError

Request processing has failed due to some unknown error, exception, or failure.

_DashboardInvalidInputError :: AsError a => Getting (First ServiceError) a ServiceError

Some part of the dashboard data is invalid.

_InvalidParameterValueException :: AsError a => Getting (First ServiceError) a ServiceError

The value of an input parameter is bad or out-of-range.

_InvalidFormatFault :: AsError a => Getting (First ServiceError) a ServiceError

Data was not syntactically valid JSON.

_MissingRequiredParameterException :: AsError a => Getting (First ServiceError) a ServiceError

An input parameter that is required is missing.

_InvalidParameterCombinationException :: AsError a => Getting (First ServiceError) a ServiceError

Parameters were used together that cannot be used together.

_ResourceNotFound :: AsError a => Getting (First ServiceError) a ServiceError

The named resource does not exist.

ComparisonOperator

data ComparisonOperator

Constructors

GreaterThanOrEqualToThreshold

GreaterThanThreshold

LessThanOrEqualToThreshold

LessThanThreshold

HistoryItemType

data HistoryItemType

Constructors

Action

ConfigurationUpdate

StateUpdate

StandardUnit

data StandardUnit

Constructors

Bits

BitsSecond

Bytes

BytesSecond

Count

CountSecond

Gigabits

GigabitsSecond

Gigabytes

GigabytesSecond

Kilobits

KilobitsSecond

Kilobytes

KilobytesSecond

Megabits

MegabitsSecond

Megabytes

MegabytesSecond

Microseconds

Milliseconds

None

Percent

Seconds

Terabits

TerabitsSecond

Terabytes

TerabytesSecond

StateValue

data StateValue

Constructors

Alarm

InsufficientData

OK

Statistic

data Statistic

Constructors

Average

Maximum

Minimum

SampleCount

Sum

AlarmHistoryItem

data AlarmHistoryItem

Represents the history of a specific alarm.

See: alarmHistoryItem smart constructor.

alarmHistoryItem :: AlarmHistoryItem

Creates a value of AlarmHistoryItem with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ahiAlarmName - The descriptive name for the alarm.

ahiHistoryItemType - The type of alarm history item.

ahiHistoryData - Data about the alarm, in JSON format.

ahiHistorySummary - A summary of the alarm history, in text format.

ahiTimestamp - The time stamp for the alarm history item.

ahiAlarmName :: Lens' AlarmHistoryItem (Maybe Text)

The descriptive name for the alarm.

ahiHistoryItemType :: Lens' AlarmHistoryItem (Maybe HistoryItemType)

The type of alarm history item.

ahiHistoryData :: Lens' AlarmHistoryItem (Maybe Text)

Data about the alarm, in JSON format.

ahiHistorySummary :: Lens' AlarmHistoryItem (Maybe Text)

A summary of the alarm history, in text format.

ahiTimestamp :: Lens' AlarmHistoryItem (Maybe UTCTime)

The time stamp for the alarm history item.

DashboardEntry

data DashboardEntry

Represents a specific dashboard.

See: dashboardEntry smart constructor.

dashboardEntry :: DashboardEntry

Creates a value of DashboardEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

deSize - The size of the dashboard, in bytes.

deDashboardName - The name of the dashboard.

deLastModified - The time stamp of when the dashboard was last modified, either by an API call or through the console. This number is expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

deDashboardARN - The Amazon Resource Name (ARN) of the dashboard.

deSize :: Lens' DashboardEntry (Maybe Integer)

The size of the dashboard, in bytes.

deDashboardName :: Lens' DashboardEntry (Maybe Text)

The name of the dashboard.

deLastModified :: Lens' DashboardEntry (Maybe UTCTime)

The time stamp of when the dashboard was last modified, either by an API call or through the console. This number is expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

deDashboardARN :: Lens' DashboardEntry (Maybe Text)

The Amazon Resource Name (ARN) of the dashboard.

DashboardValidationMessage

data DashboardValidationMessage

An error or warning for the operation.

See: dashboardValidationMessage smart constructor.

dashboardValidationMessage :: DashboardValidationMessage

Creates a value of DashboardValidationMessage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvmDataPath - The data path related to the message.

dvmMessage - A message describing the error or warning.

dvmDataPath :: Lens' DashboardValidationMessage (Maybe Text)

The data path related to the message.

dvmMessage :: Lens' DashboardValidationMessage (Maybe Text)

A message describing the error or warning.

Datapoint

data Datapoint

Encapsulates the statistical data that CloudWatch computes from metric data.

See: datapoint smart constructor.

datapoint :: Datapoint

Creates a value of Datapoint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dSampleCount - The number of metric values that contributed to the aggregate value of this data point.

dMaximum - The maximum metric value for the data point.

dAverage - The average of the metric values that correspond to the data point.

dMinimum - The minimum metric value for the data point.

dExtendedStatistics - The percentile statistic for the data point.

dSum - The sum of the metric values for the data point.

dUnit - The standard unit for the data point.

dTimestamp - The time stamp used for the data point.

dSampleCount :: Lens' Datapoint (Maybe Double)

The number of metric values that contributed to the aggregate value of this data point.

dMaximum :: Lens' Datapoint (Maybe Double)

The maximum metric value for the data point.

dAverage :: Lens' Datapoint (Maybe Double)

The average of the metric values that correspond to the data point.

dMinimum :: Lens' Datapoint (Maybe Double)

The minimum metric value for the data point.

dExtendedStatistics :: Lens' Datapoint (HashMap Text Double)

The percentile statistic for the data point.

dSum :: Lens' Datapoint (Maybe Double)

The sum of the metric values for the data point.

dUnit :: Lens' Datapoint (Maybe StandardUnit)

The standard unit for the data point.

dTimestamp :: Lens' Datapoint (Maybe UTCTime)

The time stamp used for the data point.

Dimension

data Dimension

Expands the identity of a metric.

See: dimension smart constructor.

dimension

Arguments

:: Text
dName

-> Text
dValue

-> Dimension

Creates a value of Dimension with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dName - The name of the dimension.

dValue - The value representing the dimension measurement.

dName :: Lens' Dimension Text

The name of the dimension.

dValue :: Lens' Dimension Text

The value representing the dimension measurement.

DimensionFilter

data DimensionFilter

Represents filters for a dimension.

See: dimensionFilter smart constructor.

dimensionFilter

Arguments

:: Text
dfName

-> DimensionFilter

Creates a value of DimensionFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfValue - The value of the dimension to be matched.

dfName - The dimension name to be matched.

dfValue :: Lens' DimensionFilter (Maybe Text)

The value of the dimension to be matched.

dfName :: Lens' DimensionFilter Text

The dimension name to be matched.

Metric

data Metric

Represents a specific metric.

See: metric smart constructor.

metric :: Metric

Creates a value of Metric with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mMetricName - The name of the metric.

mNamespace - The namespace of the metric.

mDimensions - The dimensions for the metric.

mMetricName :: Lens' Metric (Maybe Text)

The name of the metric.

mNamespace :: Lens' Metric (Maybe Text)

The namespace of the metric.

mDimensions :: Lens' Metric [Dimension]

The dimensions for the metric.

MetricAlarm

data MetricAlarm

Represents an alarm.

See: metricAlarm smart constructor.

metricAlarm :: MetricAlarm

Creates a value of MetricAlarm with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

maAlarmName - The name of the alarm.

maStateUpdatedTimestamp - The time stamp of the last update to the alarm state.

maTreatMissingData - Sets how this alarm is to handle missing data points. If this parameter is omitted, the default behavior of missing is used.

maPeriod - The period, in seconds, over which the statistic is applied.

maAlarmDescription - The description of the alarm.

maEvaluationPeriods - The number of periods over which data is compared to the specified threshold.

maMetricName - The name of the metric associated with the alarm.

maNamespace - The namespace of the metric associated with the alarm.

maComparisonOperator - The arithmetic operation to use when comparing the specified statistic and threshold. The specified statistic value is used as the first operand.

maOKActions - The actions to execute when this alarm transitions to the OK state from any other state. Each action is specified as an Amazon Resource Name (ARN).

maEvaluateLowSampleCountPercentile - Used only for alarms based on percentiles. If ignore , the alarm state does not change during periods with too few data points to be statistically significant. If evaluate or this parameter is not used, the alarm is always evaluated and possibly changes state no matter how many data points are available.

maStateValue - The state value for the alarm.

maThreshold - The value to compare with the specified statistic.

maAlarmConfigurationUpdatedTimestamp - The time stamp of the last update to the alarm configuration.

maActionsEnabled - Indicates whether actions should be executed during any changes to the alarm state.

maInsufficientDataActions - The actions to execute when this alarm transitions to the INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Name (ARN).

maStateReason - An explanation for the alarm state, in text format.

maStateReasonData - An explanation for the alarm state, in JSON format.

maDimensions - The dimensions for the metric associated with the alarm.

maAlarmARN - The Amazon Resource Name (ARN) of the alarm.

maAlarmActions - The actions to execute when this alarm transitions to the ALARM state from any other state. Each action is specified as an Amazon Resource Name (ARN).

maUnit - The unit of the metric associated with the alarm.

maStatistic - The statistic for the metric associated with the alarm, other than percentile. For percentile statistics, use ExtendedStatistic .

maExtendedStatistic - The percentile statistic for the metric associated with the alarm. Specify a value between p0.0 and p100.

maAlarmName :: Lens' MetricAlarm (Maybe Text)

The name of the alarm.

maStateUpdatedTimestamp :: Lens' MetricAlarm (Maybe UTCTime)

The time stamp of the last update to the alarm state.

maTreatMissingData :: Lens' MetricAlarm (Maybe Text)

Sets how this alarm is to handle missing data points. If this parameter is omitted, the default behavior of missing is used.

maPeriod :: Lens' MetricAlarm (Maybe Natural)

The period, in seconds, over which the statistic is applied.

maAlarmDescription :: Lens' MetricAlarm (Maybe Text)

The description of the alarm.

maEvaluationPeriods :: Lens' MetricAlarm (Maybe Natural)

The number of periods over which data is compared to the specified threshold.

maMetricName :: Lens' MetricAlarm (Maybe Text)

The name of the metric associated with the alarm.

maNamespace :: Lens' MetricAlarm (Maybe Text)

The namespace of the metric associated with the alarm.

maComparisonOperator :: Lens' MetricAlarm (Maybe ComparisonOperator)

The arithmetic operation to use when comparing the specified statistic and threshold. The specified statistic value is used as the first operand.

maOKActions :: Lens' MetricAlarm [Text]

The actions to execute when this alarm transitions to the OK state from any other state. Each action is specified as an Amazon Resource Name (ARN).

maEvaluateLowSampleCountPercentile :: Lens' MetricAlarm (Maybe Text)

Used only for alarms based on percentiles. If ignore , the alarm state does not change during periods with too few data points to be statistically significant. If evaluate or this parameter is not used, the alarm is always evaluated and possibly changes state no matter how many data points are available.

maStateValue :: Lens' MetricAlarm (Maybe StateValue)

The state value for the alarm.

maThreshold :: Lens' MetricAlarm (Maybe Double)

The value to compare with the specified statistic.

maAlarmConfigurationUpdatedTimestamp :: Lens' MetricAlarm (Maybe UTCTime)

The time stamp of the last update to the alarm configuration.

maActionsEnabled :: Lens' MetricAlarm (Maybe Bool)

Indicates whether actions should be executed during any changes to the alarm state.

maInsufficientDataActions :: Lens' MetricAlarm [Text]

The actions to execute when this alarm transitions to the INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Name (ARN).

maStateReason :: Lens' MetricAlarm (Maybe Text)

An explanation for the alarm state, in text format.

maStateReasonData :: Lens' MetricAlarm (Maybe Text)

An explanation for the alarm state, in JSON format.

maDimensions :: Lens' MetricAlarm [Dimension]

The dimensions for the metric associated with the alarm.

maAlarmARN :: Lens' MetricAlarm (Maybe Text)

The Amazon Resource Name (ARN) of the alarm.

maAlarmActions :: Lens' MetricAlarm [Text]

The actions to execute when this alarm transitions to the ALARM state from any other state. Each action is specified as an Amazon Resource Name (ARN).

maUnit :: Lens' MetricAlarm (Maybe StandardUnit)

The unit of the metric associated with the alarm.

maStatistic :: Lens' MetricAlarm (Maybe Statistic)

The statistic for the metric associated with the alarm, other than percentile. For percentile statistics, use ExtendedStatistic .

maExtendedStatistic :: Lens' MetricAlarm (Maybe Text)

The percentile statistic for the metric associated with the alarm. Specify a value between p0.0 and p100.

MetricDatum

data MetricDatum

Encapsulates the information sent to either create a metric or add new values to be aggregated into an existing metric.

See: metricDatum smart constructor.

metricDatum

Arguments

:: Text
mdMetricName

-> MetricDatum

Creates a value of MetricDatum with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mdValue - The value for the metric. Although the parameter accepts numbers of type Double, CloudWatch rejects values that are either too small or too large. Values must be in the range of 8.515920e-109 to 1.174271e+108 (Base 10) or 2e-360 to 2e360 (Base 2). In addition, special values (for example, NaN, +Infinity, -Infinity) are not supported.

mdStorageResolution - Valid values are 1 and 60. Setting this to 1 specifies this metric as a high-resolution metric, so that CloudWatch stores the metric with sub-minute resolution down to one second. Setting this to 60 specifies this metric as a regular-resolution metric, which CloudWatch stores at 1-minute resolution. Currently, high resolution is available only for custom metrics. For more information about high-resolution metrics, see High-Resolution Metrics in the Amazon CloudWatch User Guide . This field is optional, if you do not specify it the default of 60 is used.

mdDimensions - The dimensions associated with the metric.

mdUnit - The unit of the metric.

mdTimestamp - The time the metric data was received, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

mdStatisticValues - The statistical values for the metric.

mdMetricName - The name of the metric.

mdValue :: Lens' MetricDatum (Maybe Double)

The value for the metric. Although the parameter accepts numbers of type Double, CloudWatch rejects values that are either too small or too large. Values must be in the range of 8.515920e-109 to 1.174271e+108 (Base 10) or 2e-360 to 2e360 (Base 2). In addition, special values (for example, NaN, +Infinity, -Infinity) are not supported.

mdStorageResolution :: Lens' MetricDatum (Maybe Natural)

Valid values are 1 and 60. Setting this to 1 specifies this metric as a high-resolution metric, so that CloudWatch stores the metric with sub-minute resolution down to one second. Setting this to 60 specifies this metric as a regular-resolution metric, which CloudWatch stores at 1-minute resolution. Currently, high resolution is available only for custom metrics. For more information about high-resolution metrics, see High-Resolution Metrics in the Amazon CloudWatch User Guide . This field is optional, if you do not specify it the default of 60 is used.

mdDimensions :: Lens' MetricDatum [Dimension]

The dimensions associated with the metric.

mdUnit :: Lens' MetricDatum (Maybe StandardUnit)

The unit of the metric.

mdTimestamp :: Lens' MetricDatum (Maybe UTCTime)

The time the metric data was received, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC.

mdStatisticValues :: Lens' MetricDatum (Maybe StatisticSet)

The statistical values for the metric.

mdMetricName :: Lens' MetricDatum Text

The name of the metric.

StatisticSet

data StatisticSet

Represents a set of statistics that describes a specific metric.

See: statisticSet smart constructor.

statisticSet

Arguments

:: Double
ssSampleCount

-> Double
ssSum

-> Double
ssMinimum

-> Double
ssMaximum

-> StatisticSet

Creates a value of StatisticSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ssSampleCount - The number of samples used for the statistic set.

ssSum - The sum of values for the sample set.

ssMinimum - The minimum value of the sample set.

ssMaximum - The maximum value of the sample set.

ssSampleCount :: Lens' StatisticSet Double

The number of samples used for the statistic set.

ssSum :: Lens' StatisticSet Double

The sum of values for the sample set.

ssMinimum :: Lens' StatisticSet Double

The minimum value of the sample set.

ssMaximum :: Lens' StatisticSet Double

The maximum value of the sample set.