Network.AWS.CloudWatch.Types

Service

data CloudWatch

Version 2010-08-01 of the Amazon CloudWatch service.

Error

data RESTError :: *

XML

ns :: Text

StatisticSet

data StatisticSet

statisticSet

Arguments

:: Double
ssSampleCount

-> Double
ssSum

-> Double
ssMinimum

-> Double
ssMaximum

-> StatisticSet

StatisticSet constructor.

The fields accessible through corresponding lenses are:

ssMaximum :: Double

ssMinimum :: Double

ssSampleCount :: Double

ssSum :: Double

ssMaximum :: Lens' StatisticSet Double

The maximum value of the sample set.

ssMinimum :: Lens' StatisticSet Double

The minimum value of the sample set.

ssSampleCount :: Lens' StatisticSet Double

The number of samples used for the statistic set.

ssSum :: Lens' StatisticSet Double

The sum of values for the sample set.

MetricAlarm

data MetricAlarm

metricAlarm :: MetricAlarm

MetricAlarm constructor.

The fields accessible through corresponding lenses are:

maActionsEnabled :: Maybe Bool

maAlarmActions :: [Text]

maAlarmArn :: Maybe Text

maAlarmConfigurationUpdatedTimestamp :: Maybe UTCTime

maAlarmDescription :: Maybe Text

maAlarmName :: Maybe Text

maComparisonOperator :: Maybe ComparisonOperator

maDimensions :: [Dimension]

maEvaluationPeriods :: Maybe Natural

maInsufficientDataActions :: [Text]

maMetricName :: Maybe Text

maNamespace :: Maybe Text

maOKActions :: [Text]

maPeriod :: Maybe Natural

maStateReason :: Maybe Text

maStateReasonData :: Maybe Text

maStateUpdatedTimestamp :: Maybe UTCTime

maStateValue :: Maybe StateValue

maStatistic :: Maybe Statistic

maThreshold :: Maybe Double

maUnit :: Maybe StandardUnit

maActionsEnabled :: Lens' MetricAlarm (Maybe Bool)

Indicates whether actions should be executed during any changes to the alarm's state.

maAlarmActions :: Lens' MetricAlarm [Text]

The list of actions to execute when this alarm transitions into an ALARM state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only actions supported are publishing to an Amazon SNS topic and triggering an Auto Scaling policy.

maAlarmArn :: Lens' MetricAlarm (Maybe Text)

The Amazon Resource Name (ARN) of the alarm.

maAlarmConfigurationUpdatedTimestamp :: Lens' MetricAlarm (Maybe UTCTime)

The time stamp of the last update to the alarm configuration. Amazon CloudWatch uses Coordinated Universal Time (UTC) when returning time stamps, which do not accommodate seasonal adjustments such as daylight savings time. For more information, see Time stamps in the Amazon CloudWatch Developer Guide .

maAlarmDescription :: Lens' MetricAlarm (Maybe Text)

The description for the alarm.

maAlarmName :: Lens' MetricAlarm (Maybe Text)

The name of the alarm.

maComparisonOperator :: Lens' MetricAlarm (Maybe ComparisonOperator)

The arithmetic operation to use when comparing the specified Statistic and Threshold. The specified Statistic value is used as the first operand.

maDimensions :: Lens' MetricAlarm [Dimension]

The list of dimensions associated with the alarm's associated metric.

maEvaluationPeriods :: Lens' MetricAlarm (Maybe Natural)

The number of periods over which data is compared to the specified threshold.

maInsufficientDataActions :: Lens' MetricAlarm [Text]

The list of actions to execute when this alarm transitions into an INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only actions supported are publishing to an Amazon SNS topic or triggering an Auto Scaling policy.

The current WSDL lists this attribute as UnknownActions.

maMetricName :: Lens' MetricAlarm (Maybe Text)

The name of the alarm's metric.

maNamespace :: Lens' MetricAlarm (Maybe Text)

The namespace of alarm's associated metric.

maOKActions :: Lens' MetricAlarm [Text]

The list of actions to execute when this alarm transitions into an OK state from any other state. Each action is specified as an Amazon Resource Number (ARN). Currently the only actions supported are publishing to an Amazon SNS topic and triggering an Auto Scaling policy.

maPeriod :: Lens' MetricAlarm (Maybe Natural)

The period in seconds over which the statistic is applied.

maStateReason :: Lens' MetricAlarm (Maybe Text)

A human-readable explanation for the alarm's state.

maStateReasonData :: Lens' MetricAlarm (Maybe Text)

An explanation for the alarm's state in machine-readable JSON format

maStateUpdatedTimestamp :: Lens' MetricAlarm (Maybe UTCTime)

The time stamp of the last update to the alarm's state. Amazon CloudWatch uses Coordinated Universal Time (UTC) when returning time stamps, which do not accommodate seasonal adjustments such as daylight savings time. For more information, see Time stamps in the Amazon CloudWatch Developer Guide.

maStateValue :: Lens' MetricAlarm (Maybe StateValue)

The state value for the alarm.

maStatistic :: Lens' MetricAlarm (Maybe Statistic)

The statistic to apply to the alarm's associated metric.

maThreshold :: Lens' MetricAlarm (Maybe Double)

The value against which the specified statistic is compared.

maUnit :: Lens' MetricAlarm (Maybe StandardUnit)

The unit of the alarm's associated metric.

HistoryItemType

data HistoryItemType

Constructors

Action
Action

ConfigurationUpdate
ConfigurationUpdate

StateUpdate
StateUpdate

MetricDatum

data MetricDatum

metricDatum

Arguments

:: Text
mdMetricName

-> MetricDatum

MetricDatum constructor.

The fields accessible through corresponding lenses are:

mdDimensions :: [Dimension]

mdMetricName :: Text

mdStatisticValues :: Maybe StatisticSet

mdTimestamp :: Maybe UTCTime

mdUnit :: Maybe StandardUnit

mdValue :: Maybe Double

mdDimensions :: Lens' MetricDatum [Dimension]

A list of dimensions associated with the metric. Note, when using the Dimensions value in a query, you need to append .member.N to it (e.g., Dimensions.member.N).

mdMetricName :: Lens' MetricDatum Text

The name of the metric.

mdStatisticValues :: Lens' MetricDatum (Maybe StatisticSet)

A set of statistical values describing the metric.

mdTimestamp :: Lens' MetricDatum (Maybe UTCTime)

The time stamp used for the metric. If not specified, the default value is set to the time the metric data was received. Amazon CloudWatch uses Coordinated Universal Time (UTC) when returning time stamps, which do not accommodate seasonal adjustments such as daylight savings time. For more information, see Time stamps in the Amazon CloudWatch Developer Guide.

mdUnit :: Lens' MetricDatum (Maybe StandardUnit)

The unit of the metric.

mdValue :: Lens' MetricDatum (Maybe Double)

The value for the metric.

Although the Value parameter accepts numbers of type Double, Amazon CloudWatch truncates values with very large exponents. Values with base-10 exponents greater than 126 (1 x 10^126) are truncated. Likewise, values with base-10 exponents less than -130 (1 x 10^-130) are also truncated.

StandardUnit

data StandardUnit

Constructors

Bits
Bits

BitsSecond
Bits/Second

Bytes
Bytes

BytesSecond
Bytes/Second

Count
Count

CountSecond
Count/Second

Gigabits
Gigabits

GigabitsSecond
Gigabits/Second

Gigabytes
Gigabytes

GigabytesSecond
Gigabytes/Second

Kilobits
Kilobits

KilobitsSecond
Kilobits/Second

Kilobytes
Kilobytes

KilobytesSecond
Kilobytes/Second

Megabits
Megabits

MegabitsSecond
Megabits/Second

Megabytes
Megabytes

MegabytesSecond
Megabytes/Second

Microseconds
Microseconds

Milliseconds
Milliseconds

None
None

Percent
Percent

Seconds
Seconds

Terabits
Terabits

TerabitsSecond
Terabits/Second

Terabytes
Terabytes

TerabytesSecond
Terabytes/Second

Dimension

data Dimension

dimension

Arguments

:: Text
dName

-> Text
dValue

-> Dimension

Dimension constructor.

The fields accessible through corresponding lenses are:

dName :: Text

dValue :: Text

dName :: Lens' Dimension Text

The name of the dimension.

dValue :: Lens' Dimension Text

The value representing the dimension measurement

ComparisonOperator

data ComparisonOperator

Constructors

GreaterThanOrEqualToThreshold
GreaterThanOrEqualToThreshold

GreaterThanThreshold
GreaterThanThreshold

LessThanOrEqualToThreshold
LessThanOrEqualToThreshold

LessThanThreshold
LessThanThreshold

AlarmHistoryItem

data AlarmHistoryItem

alarmHistoryItem :: AlarmHistoryItem

AlarmHistoryItem constructor.

The fields accessible through corresponding lenses are:

ahiAlarmName :: Maybe Text

ahiHistoryData :: Maybe Text

ahiHistoryItemType :: Maybe HistoryItemType

ahiHistorySummary :: Maybe Text

ahiTimestamp :: Maybe UTCTime

ahiAlarmName :: Lens' AlarmHistoryItem (Maybe Text)

The descriptive name for the alarm.

ahiHistoryData :: Lens' AlarmHistoryItem (Maybe Text)

Machine-readable data about the alarm in JSON format.

ahiHistoryItemType :: Lens' AlarmHistoryItem (Maybe HistoryItemType)

The type of alarm history item.

ahiHistorySummary :: Lens' AlarmHistoryItem (Maybe Text)

A human-readable summary of the alarm history.

ahiTimestamp :: Lens' AlarmHistoryItem (Maybe UTCTime)

The time stamp for the alarm history item. Amazon CloudWatch uses Coordinated Universal Time (UTC) when returning time stamps, which do not accommodate seasonal adjustments such as daylight savings time. For more information, see Time stamps in the Amazon CloudWatch Developer Guide.

Metric

data Metric

metric :: Metric

Metric constructor.

The fields accessible through corresponding lenses are:

mDimensions :: [Dimension]

mMetricName :: Maybe Text

mNamespace :: Maybe Text

mDimensions :: Lens' Metric [Dimension]

A list of dimensions associated with the metric.

mMetricName :: Lens' Metric (Maybe Text)

The name of the metric.

mNamespace :: Lens' Metric (Maybe Text)

The namespace of the metric.

StateValue

data StateValue

Constructors

Alarm
ALARM

InsufficientData
INSUFFICIENT_DATA

Ok
OK

Datapoint

data Datapoint

datapoint :: Datapoint

Datapoint constructor.

The fields accessible through corresponding lenses are:

dAverage :: Maybe Double

dMaximum :: Maybe Double

dMinimum :: Maybe Double

dSampleCount :: Maybe Double

dSum :: Maybe Double

dTimestamp :: Maybe UTCTime

dUnit :: Maybe StandardUnit

dAverage :: Lens' Datapoint (Maybe Double)

The average of metric values that correspond to the datapoint.

dMaximum :: Lens' Datapoint (Maybe Double)

The maximum of the metric value used for the datapoint.

dMinimum :: Lens' Datapoint (Maybe Double)

The minimum metric value used for the datapoint.

dSampleCount :: Lens' Datapoint (Maybe Double)

The number of metric values that contributed to the aggregate value of this datapoint.

dSum :: Lens' Datapoint (Maybe Double)

The sum of metric values used for the datapoint.

dTimestamp :: Lens' Datapoint (Maybe UTCTime)

The time stamp used for the datapoint. Amazon CloudWatch uses Coordinated Universal Time (UTC) when returning time stamps, which do not accommodate seasonal adjustments such as daylight savings time. For more information, see Time stamps in the Amazon CloudWatch Developer Guide.

dUnit :: Lens' Datapoint (Maybe StandardUnit)

The standard unit used for the datapoint.

DimensionFilter

data DimensionFilter

dimensionFilter

Arguments

:: Text
dfName

-> DimensionFilter

DimensionFilter constructor.

The fields accessible through corresponding lenses are:

dfName :: Text

dfValue :: Maybe Text

dfName :: Lens' DimensionFilter Text

The dimension name to be matched.

dfValue :: Lens' DimensionFilter (Maybe Text)

The value of the dimension to be matched.

Statistic

data Statistic

Constructors

Average
Average

Maximum
Maximum

Minimum
Minimum

SampleCount
SampleCount

Sum
Sum